@@ -32,7 +32,7 @@ var (
 // AccessManagerMetaData contains all meta data concerning the AccessManager contract.
 var AccessManagerMetaData = &bind.MetaData{
 	ABI: "[{\"inputs\":[{\"internalType\":\"address\",\"name\":\"storageAddress\",\"type\":\"address\"}],\"stateMutability\":\"nonpayable\",\"type\":\"constructor\"},{\"inputs\":[],\"name\":\"BucketNotFound\",\"type\":\"error\"},{\"inputs\":[],\"name\":\"FileDoesNotExist\",\"type\":\"error\"},{\"inputs\":[],\"name\":\"InvalidAddress\",\"type\":\"error\"},{\"inputs\":[],\"name\":\"NoDelegatedAccess\",\"type\":\"error\"},{\"inputs\":[],\"name\":\"NoPolicy\",\"type\":\"error\"},{\"inputs\":[],\"name\":\"NotBucketOwner\",\"type\":\"error\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"internalType\":\"bytes32\",\"name\":\"bucketId\",\"type\":\"bytes32\"},{\"indexed\":true,\"internalType\":\"address\",\"name\":\"delegate\",\"type\":\"address\"}],\"name\":\"AccessDelegated\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"internalType\":\"bytes32\",\"name\":\"bucketId\",\"type\":\"bytes32\"},{\"indexed\":true,\"internalType\":\"address\",\"name\":\"delegate\",\"type\":\"address\"}],\"name\":\"AccessRemoved\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"internalType\":\"bytes32\",\"name\":\"fileId\",\"type\":\"bytes32\"},{\"indexed\":true,\"internalType\":\"address\",\"name\":\"policyContract\",\"type\":\"address\"}],\"name\":\"PolicyChanged\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"internalType\":\"bytes32\",\"name\":\"fileId\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"bool\",\"name\":\"isPublic\",\"type\":\"bool\"}],\"name\":\"PublicAccessChanged\",\"type\":\"event\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"fileId\",\"type\":\"bytes32\"},{\"internalType\":\"bool\",\"name\":\"isPublic\",\"type\":\"bool\"}],\"name\":\"changePublicAccess\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"bucketId\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"delegated\",\"type\":\"address\"}],\"name\":\"delegateAccess\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"fileId\",\"type\":\"bytes32\"}],\"name\":\"getFileAccessInfo\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"},{\"internalType\":\"bool\",\"name\":\"\",\"type\":\"bool\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"fileId\",\"type\":\"bytes32\"}],\"name\":\"getPolicy\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"fileId\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"user\",\"type\":\"address\"},{\"internalType\":\"bytes\",\"name\":\"data\",\"type\":\"bytes\"}],\"name\":\"getValidateAccess\",\"outputs\":[{\"internalType\":\"bool\",\"name\":\"\",\"type\":\"bool\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"bucketId\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"user\",\"type\":\"address\"},{\"internalType\":\"bytes\",\"name\":\"data\",\"type\":\"bytes\"}],\"name\":\"getValidateAccessToBucket\",\"outputs\":[{\"internalType\":\"bool\",\"name\":\"\",\"type\":\"bool\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"bucketId\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"user\",\"type\":\"address\"}],\"name\":\"isBucketOwnerOrDelegate\",\"outputs\":[{\"internalType\":\"bool\",\"name\":\"\",\"type\":\"bool\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"bucketId\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"delegated\",\"type\":\"address\"}],\"name\":\"removeAccess\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"bucketId\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"policyContract\",\"type\":\"address\"}],\"name\":\"setBucketPolicy\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"fileId\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"policyContract\",\"type\":\"address\"}],\"name\":\"setFilePolicy\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"address\",\"name\":\"_storageAddress\",\"type\":\"address\"}],\"name\":\"setStorageContract\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[],\"name\":\"storageContract\",\"outputs\":[{\"internalType\":\"contractIStorage\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"}]",
-	Bin: "0x6080604052348015600e575f5ffd5b506040516113ed3803806113ed833981016040819052602b91604e565b5f80546001600160a01b0319166001600160a01b03929092169190911790556079565b5f60208284031215605d575f5ffd5b81516001600160a01b03811681146072575f5ffd5b9392505050565b611367806100865f395ff3fe608060405234801561000f575f5ffd5b50600436106100b1575f3560e01c8063a3f685f91161006e578063a3f685f91461019d578063a9b61d4d146101b0578063c05d0d6c146101c3578063dc38b0a2146101d6578063e124bdd914610205578063e13212f814610218575f5ffd5b806311ce0267146100b557806338f5f5b2146100e45780633a82c912146100f957806344078802146101545780635e2a22de14610177578063a20e18df1461018a575b5f5ffd5b5f546100c7906001600160a01b031681565b6040516001600160a01b0390911681526020015b60405180910390f35b6100f76100f2366004610c2d565b61022b565b005b610135610107366004610c5b565b5f908152600160209081526040808320546003909252909120546001600160a01b039091169160ff90911690565b604080516001600160a01b0390931683529015156020830152016100db565b610167610162366004610c86565b610325565b60405190151581526020016100db565b6100f7610185366004610d09565b610461565b610167610198366004610c86565b610530565b6100c76101ab366004610c5b565b61062d565b6100f76101be366004610d09565b610719565b6100f76101d1366004610d09565b6107a5565b6100f76101e4366004610d2c565b5f80546001600160a01b0319166001600160a01b0392909216919091179055565b610167610213366004610d09565b610808565b6100f7610226366004610d09565b610966565b5f8054604051637d7602a160e11b8152600481018590528492916001600160a01b03169063faec0542906024015f60405180830381865afa158015610272573d5f5f3e3d5ffd5b505050506040513d5f823e601f3d908101601f191682016040526102999190810190610fe0565b60408101519091505f036102c057604051631de5d86f60e11b815260040160405180910390fd5b6102cd8160400151610a62565b5f84815260036020908152604091829020805460ff1916861515908117909155915191825285917f9500a58cfb37cef230929cd9f25ce92c41374416f23b1825232d0905a7e73d5a910160405180910390a250505050565b5f8481526003602052604081205460ff16806103b357505f546040516306d1d7bd60e21b8152600481018790526001600160a01b03868116921690631b475ef490602401602060405180830381865afa158015610384573d5f5f3e3d5ffd5b505050506040513d601f19601f820116820180604052508101906103a891906110d1565b6001600160a01b0316145b156103c057506001610459565b5f6103ca8661062d565b90506001600160a01b0381166103e3575f915050610459565b604051631704429d60e21b815281906001600160a01b03821690635c110a7490610415908990899089906004016110ec565b602060405180830381865afa158015610430573d5f5f3e3d5ffd5b505050506040513d601f19601f82011682018060405250810190610454919061112b565b925050505b949350505050565b8161046b81610a62565b6001600160a01b0382166104925760405163e6c4247b60e01b815260040160405180910390fd5b5f8381526004602090815260408083206001600160a01b038616845290915281205460ff16151590036104d857604051635571f9bf60e11b815260040160405180910390fd5b5f8381526004602090815260408083206001600160a01b0386168085529252808320805460ff1916905551909185917f907b4ea595c8ab4b33e5ce3b70882b068119d80536a1121f37c7ab73104e2b739190a3505050565b5f5f61053b86610ac6565b60608101519091506001600160a01b031633148061057157505f86815260046020908152604080832033845290915290205460ff165b15610580576001915050610459565b61058986610ac6565b505f868152600260205260409020546001600160a01b0316806105b0575f92505050610459565b604051631704429d60e21b815281906001600160a01b03821690635c110a74906105e2908a908a908a906004016110ec565b602060405180830381865afa1580156105fd573d5f5f3e3d5ffd5b505050506040513d601f19601f82011682018060405250810190610621919061112b565b98975050505050505050565b5f818152600160205260408120546001600160a01b03166106fe575f8054604051637d7602a160e11b8152600481018590526001600160a01b039091169063faec0542906024015f60405180830381865afa15801561068e573d5f5f3e3d5ffd5b505050506040513d5f823e601f3d908101601f191682016040526106b59190810190610fe0565b6040808201515f908152600260205220549091506001600160a01b03166106de57505f92915050565b6040908101515f908152600260205220546001600160a01b031692915050565b505f908152600160205260409020546001600160a01b031690565b8161072381610a62565b6001600160a01b03821661074a5760405163e6c4247b60e01b815260040160405180910390fd5b5f8381526004602090815260408083206001600160a01b0386168085529252808320805460ff1916600117905551909185917f4c71ed01d4ddaea8f2390c2feb6be34befe3571daf558f29632daebe2a1724339190a3505050565b816107af81610a62565b5f8381526002602052604080822080546001600160a01b0319166001600160a01b0386169081179091559051909185917fee78f4c2af2887839fdcba441de8e4d2d1b117b89ce7cc7a7f0a952871cc87d29190a3505050565b6040805160018082528183019092525f918291906020808301908036833701905050905083815f8151811061083f5761083f611146565b60209081029190910101525f80546040516335bdb71160e01b81526001600160a01b03909116906335bdb7119061087a90859060040161115a565b5f60405180830381865afa158015610894573d5f5f3e3d5ffd5b505050506040513d5f823e601f3d908101601f191682016040526108bb919081019061119c565b905080515f14806108e957505f5f1b815f815181106108dc576108dc611146565b60200260200101515f0151145b156108f8575f92505050610960565b5f815f8151811061090b5761090b611146565b6020026020010151606001519050846001600160a01b0316816001600160a01b0316148061095a57505f8681526004602090815260408083206001600160a01b038916845290915290205460ff165b93505050505b92915050565b5f8054604051637d7602a160e11b8152600481018590528492916001600160a01b03169063faec0542906024015f60405180830381865afa1580156109ad573d5f5f3e3d5ffd5b505050506040513d5f823e601f3d908101601f191682016040526109d49190810190610fe0565b60408101519091505f036109fb57604051631de5d86f60e11b815260040160405180910390fd5b610a088160400151610a62565b5f8481526001602052604080822080546001600160a01b0319166001600160a01b0387169081179091559051909186917fee78f4c2af2887839fdcba441de8e4d2d1b117b89ce7cc7a7f0a952871cc87d29190a350505050565b5f610a6c82610ac6565b60608101519091506001600160a01b03163314801590610aa557505f82815260046020908152604080832033845290915290205460ff16155b15610ac25760405162d6b18f60e41b815260040160405180910390fd5b5050565b610afe6040518060a001604052805f8152602001606081526020015f81526020015f6001600160a01b03168152602001606081525090565b6040805160018082528183019092525f916020808301908036833701905050905082815f81518110610b3257610b32611146565b60209081029190910101525f80546040516335bdb71160e01b81526001600160a01b03909116906335bdb71190610b6d90859060040161115a565b5f60405180830381865afa158015610b87573d5f5f3e3d5ffd5b505050506040513d5f823e601f3d908101601f19168201604052610bae919081019061119c565b905080515f1480610bdc57505f5f1b815f81518110610bcf57610bcf611146565b60200260200101515f0151145b15610bfa5760405163c4c1a0c560e01b815260040160405180910390fd5b805f81518110610c0c57610c0c611146565b602002602001015192505050919050565b8015158114610c2a575f5ffd5b50565b5f5f60408385031215610c3e575f5ffd5b823591506020830135610c5081610c1d565b809150509250929050565b5f60208284031215610c6b575f5ffd5b5035919050565b6001600160a01b0381168114610c2a575f5ffd5b5f5f5f5f60608587031215610c99575f5ffd5b843593506020850135610cab81610c72565b925060408501356001600160401b03811115610cc5575f5ffd5b8501601f81018713610cd5575f5ffd5b80356001600160401b03811115610cea575f5ffd5b876020828401011115610cfb575f5ffd5b949793965060200194505050565b5f5f60408385031215610d1a575f5ffd5b823591506020830135610c5081610c72565b5f60208284031215610d3c575f5ffd5b8135610d4781610c72565b9392505050565b634e487b7160e01b5f52604160045260245ffd5b604080519081016001600160401b0381118282101715610d8457610d84610d4e565b60405290565b60405161010081016001600160401b0381118282101715610d8457610d84610d4e565b60405160a081016001600160401b0381118282101715610d8457610d84610d4e565b604051601f8201601f191681016001600160401b0381118282101715610df757610df7610d4e565b604052919050565b5f82601f830112610e0e575f5ffd5b8151602083015f5f6001600160401b03841115610e2d57610e2d610d4e565b50601f8301601f1916602001610e4281610dcf565b915050828152858383011115610e56575f5ffd5b8282602083015e5f92810160200192909252509392505050565b5f6001600160401b03821115610e8857610e88610d4e565b5060051b60200190565b5f82601f830112610ea1575f5ffd5b8151610eb4610eaf82610e70565b610dcf565b8082825260208201915060208360051b860101925085831115610ed5575f5ffd5b602085015b83811015610ef2578051835260209283019201610eda565b5095945050505050565b5f60408284031215610f0c575f5ffd5b610f14610d62565b905081516001600160401b03811115610f2b575f5ffd5b8201601f81018413610f3b575f5ffd5b8051610f49610eaf82610e70565b8082825260208201915060208360051b850101925086831115610f6a575f5ffd5b602084015b83811015610faa5780516001600160401b03811115610f8c575f5ffd5b610f9b89602083890101610dff565b84525060209283019201610f6f565b50845250505060208201516001600160401b03811115610fc8575f5ffd5b610fd484828501610e92565b60208301525092915050565b5f60208284031215610ff0575f5ffd5b81516001600160401b03811115611005575f5ffd5b82016101008185031215611017575f5ffd5b61101f610d8a565b8151815260208201516001600160401b0381111561103b575f5ffd5b61104786828501610dff565b6020830152506040828101519082015260608201516001600160401b0381111561106f575f5ffd5b61107b86828501610dff565b6060830152506080828101519082015260a0808301519082015260c0808301519082015260e08201516001600160401b038111156110b7575f5ffd5b6110c386828501610efc565b60e083015250949350505050565b5f602082840312156110e1575f5ffd5b8151610d4781610c72565b6001600160a01b03841681526040602082018190528101829052818360608301375f818301606090810191909152601f909201601f1916010192915050565b5f6020828403121561113b575f5ffd5b8151610d4781610c1d565b634e487b7160e01b5f52603260045260245ffd5b602080825282518282018190525f918401906040840190835b81811015611191578351835260209384019390920191600101611173565b509095945050505050565b5f602082840312156111ac575f5ffd5b81516001600160401b038111156111c1575f5ffd5b8201601f810184136111d1575f5ffd5b80516111df610eaf82610e70565b8082825260208201915060208360051b850101925086831115611200575f5ffd5b602084015b838110156113265780516001600160401b03811115611222575f5ffd5b850160a0818a03601f19011215611237575f5ffd5b61123f610dad565b6020820151815260408201516001600160401b0381111561125e575f5ffd5b61126d8b602083860101610dff565b60208301525060608201516040820152608082015161128b81610c72565b606082015260a08201516001600160401b038111156112a8575f5ffd5b60208184010192505089601f8301126112bf575f5ffd5b81516112cd610eaf82610e70565b8082825260208201915060208360051b86010192508c8311156112ee575f5ffd5b6020850194505b828510156113105784518252602094850194909101906112f5565b6080840152505084525060209283019201611205565b50969550505050505056fea2646970667358221220ba5c94e1acfaa11d994296f0628916efc93db984290cf7d8b09ffb7127b3619964736f6c634300081c0033",
+	Bin: "0x6080604052348015600e575f5ffd5b50604051611483380380611483833981016040819052602b91604e565b5f80546001600160a01b0319166001600160a01b03929092169190911790556079565b5f60208284031215605d575f5ffd5b81516001600160a01b03811681146072575f5ffd5b9392505050565b6113fd806100865f395ff3fe608060405234801561000f575f5ffd5b50600436106100b1575f3560e01c8063a3f685f91161006e578063a3f685f91461019d578063a9b61d4d146101b0578063c05d0d6c146101c3578063dc38b0a2146101d6578063e124bdd914610205578063e13212f814610218575f5ffd5b806311ce0267146100b557806338f5f5b2146100e45780633a82c912146100f957806344078802146101545780635e2a22de14610177578063a20e18df1461018a575b5f5ffd5b5f546100c7906001600160a01b031681565b6040516001600160a01b0390911681526020015b60405180910390f35b6100f76100f2366004610c2d565b61022b565b005b610135610107366004610c5b565b5f908152600160209081526040808320546003909252909120546001600160a01b039091169160ff90911690565b604080516001600160a01b0390931683529015156020830152016100db565b610167610162366004610c86565b610325565b60405190151581526020016100db565b6100f7610185366004610d09565b610461565b610167610198366004610c86565b610530565b6100c76101ab366004610c5b565b61062d565b6100f76101be366004610d09565b610719565b6100f76101d1366004610d09565b6107a5565b6100f76101e4366004610d2c565b5f80546001600160a01b0319166001600160a01b0392909216919091179055565b610167610213366004610d09565b610808565b6100f7610226366004610d09565b610966565b5f8054604051637d7602a160e11b8152600481018590528492916001600160a01b03169063faec0542906024015f60405180830381865afa158015610272573d5f5f3e3d5ffd5b505050506040513d5f823e601f3d908101601f191682016040526102999190810190611076565b60408101519091505f036102c057604051631de5d86f60e11b815260040160405180910390fd5b6102cd8160400151610a62565b5f84815260036020908152604091829020805460ff1916861515908117909155915191825285917f9500a58cfb37cef230929cd9f25ce92c41374416f23b1825232d0905a7e73d5a910160405180910390a250505050565b5f8481526003602052604081205460ff16806103b357505f546040516306d1d7bd60e21b8152600481018790526001600160a01b03868116921690631b475ef490602401602060405180830381865afa158015610384573d5f5f3e3d5ffd5b505050506040513d601f19601f820116820180604052508101906103a89190611167565b6001600160a01b0316145b156103c057506001610459565b5f6103ca8661062d565b90506001600160a01b0381166103e3575f915050610459565b604051631704429d60e21b815281906001600160a01b03821690635c110a749061041590899089908990600401611182565b602060405180830381865afa158015610430573d5f5f3e3d5ffd5b505050506040513d601f19601f8201168201806040525081019061045491906111c1565b925050505b949350505050565b8161046b81610a62565b6001600160a01b0382166104925760405163e6c4247b60e01b815260040160405180910390fd5b5f8381526004602090815260408083206001600160a01b038616845290915281205460ff16151590036104d857604051635571f9bf60e11b815260040160405180910390fd5b5f8381526004602090815260408083206001600160a01b0386168085529252808320805460ff1916905551909185917f907b4ea595c8ab4b33e5ce3b70882b068119d80536a1121f37c7ab73104e2b739190a3505050565b5f5f61053b86610ac6565b60608101519091506001600160a01b031633148061057157505f86815260046020908152604080832033845290915290205460ff165b15610580576001915050610459565b61058986610ac6565b505f868152600260205260409020546001600160a01b0316806105b0575f92505050610459565b604051631704429d60e21b815281906001600160a01b03821690635c110a74906105e2908a908a908a90600401611182565b602060405180830381865afa1580156105fd573d5f5f3e3d5ffd5b505050506040513d601f19601f8201168201806040525081019061062191906111c1565b98975050505050505050565b5f818152600160205260408120546001600160a01b03166106fe575f8054604051637d7602a160e11b8152600481018590526001600160a01b039091169063faec0542906024015f60405180830381865afa15801561068e573d5f5f3e3d5ffd5b505050506040513d5f823e601f3d908101601f191682016040526106b59190810190611076565b6040808201515f908152600260205220549091506001600160a01b03166106de57505f92915050565b6040908101515f908152600260205220546001600160a01b031692915050565b505f908152600160205260409020546001600160a01b031690565b8161072381610a62565b6001600160a01b03821661074a5760405163e6c4247b60e01b815260040160405180910390fd5b5f8381526004602090815260408083206001600160a01b0386168085529252808320805460ff1916600117905551909185917f4c71ed01d4ddaea8f2390c2feb6be34befe3571daf558f29632daebe2a1724339190a3505050565b816107af81610a62565b5f8381526002602052604080822080546001600160a01b0319166001600160a01b0386169081179091559051909185917fee78f4c2af2887839fdcba441de8e4d2d1b117b89ce7cc7a7f0a952871cc87d29190a3505050565b6040805160018082528183019092525f918291906020808301908036833701905050905083815f8151811061083f5761083f6111dc565b60209081029190910101525f80546040516335bdb71160e01b81526001600160a01b03909116906335bdb7119061087a9085906004016111f0565b5f60405180830381865afa158015610894573d5f5f3e3d5ffd5b505050506040513d5f823e601f3d908101601f191682016040526108bb9190810190611232565b905080515f14806108e957505f5f1b815f815181106108dc576108dc6111dc565b60200260200101515f0151145b156108f8575f92505050610960565b5f815f8151811061090b5761090b6111dc565b6020026020010151606001519050846001600160a01b0316816001600160a01b0316148061095a57505f8681526004602090815260408083206001600160a01b038916845290915290205460ff165b93505050505b92915050565b5f8054604051637d7602a160e11b8152600481018590528492916001600160a01b03169063faec0542906024015f60405180830381865afa1580156109ad573d5f5f3e3d5ffd5b505050506040513d5f823e601f3d908101601f191682016040526109d49190810190611076565b60408101519091505f036109fb57604051631de5d86f60e11b815260040160405180910390fd5b610a088160400151610a62565b5f8481526001602052604080822080546001600160a01b0319166001600160a01b0387169081179091559051909186917fee78f4c2af2887839fdcba441de8e4d2d1b117b89ce7cc7a7f0a952871cc87d29190a350505050565b5f610a6c82610ac6565b60608101519091506001600160a01b03163314801590610aa557505f82815260046020908152604080832033845290915290205460ff16155b15610ac25760405162d6b18f60e41b815260040160405180910390fd5b5050565b610afe6040518060a001604052805f8152602001606081526020015f81526020015f6001600160a01b03168152602001606081525090565b6040805160018082528183019092525f916020808301908036833701905050905082815f81518110610b3257610b326111dc565b60209081029190910101525f80546040516335bdb71160e01b81526001600160a01b03909116906335bdb71190610b6d9085906004016111f0565b5f60405180830381865afa158015610b87573d5f5f3e3d5ffd5b505050506040513d5f823e601f3d908101601f19168201604052610bae9190810190611232565b905080515f1480610bdc57505f5f1b815f81518110610bcf57610bcf6111dc565b60200260200101515f0151145b15610bfa5760405163c4c1a0c560e01b815260040160405180910390fd5b805f81518110610c0c57610c0c6111dc565b602002602001015192505050919050565b8015158114610c2a575f5ffd5b50565b5f5f60408385031215610c3e575f5ffd5b823591506020830135610c5081610c1d565b809150509250929050565b5f60208284031215610c6b575f5ffd5b5035919050565b6001600160a01b0381168114610c2a575f5ffd5b5f5f5f5f60608587031215610c99575f5ffd5b843593506020850135610cab81610c72565b925060408501356001600160401b03811115610cc5575f5ffd5b8501601f81018713610cd5575f5ffd5b80356001600160401b03811115610cea575f5ffd5b876020828401011115610cfb575f5ffd5b949793965060200194505050565b5f5f60408385031215610d1a575f5ffd5b823591506020830135610c5081610c72565b5f60208284031215610d3c575f5ffd5b8135610d4781610c72565b9392505050565b634e487b7160e01b5f52604160045260245ffd5b604051606081016001600160401b0381118282101715610d8457610d84610d4e565b60405290565b60405161010081016001600160401b0381118282101715610d8457610d84610d4e565b60405160a081016001600160401b0381118282101715610d8457610d84610d4e565b604051601f8201601f191681016001600160401b0381118282101715610df757610df7610d4e565b604052919050565b5f82601f830112610e0e575f5ffd5b8151602083015f5f6001600160401b03841115610e2d57610e2d610d4e565b50601f8301601f1916602001610e4281610dcf565b915050828152858383011115610e56575f5ffd5b8282602083015e5f92810160200192909252509392505050565b5f6001600160401b03821115610e8857610e88610d4e565b5060051b60200190565b5f82601f830112610ea1575f5ffd5b8151610eb4610eaf82610e70565b610dcf565b8082825260208201915060208360051b860101925085831115610ed5575f5ffd5b602085015b83811015610ef2578051835260209283019201610eda565b5095945050505050565b5f82601f830112610f0b575f5ffd5b8151610f19610eaf82610e70565b8082825260208201915060208360051b860101925085831115610f3a575f5ffd5b602085015b83811015610ef257805163ffffffff81168114610f5a575f5ffd5b835260209283019201610f3f565b5f60608284031215610f78575f5ffd5b610f80610d62565b905081516001600160401b03811115610f97575f5ffd5b8201601f81018413610fa7575f5ffd5b8051610fb5610eaf82610e70565b8082825260208201915060208360051b850101925086831115610fd6575f5ffd5b602084015b838110156110165780516001600160401b03811115610ff8575f5ffd5b61100789602083890101610dff565b84525060209283019201610fdb565b50845250505060208201516001600160401b03811115611034575f5ffd5b61104084828501610e92565b60208301525060408201516001600160401b0381111561105e575f5ffd5b61106a84828501610efc565b60408301525092915050565b5f60208284031215611086575f5ffd5b81516001600160401b0381111561109b575f5ffd5b820161010081850312156110ad575f5ffd5b6110b5610d8a565b8151815260208201516001600160401b038111156110d1575f5ffd5b6110dd86828501610dff565b6020830152506040828101519082015260608201516001600160401b03811115611105575f5ffd5b61111186828501610dff565b6060830152506080828101519082015260a0808301519082015260c0808301519082015260e08201516001600160401b0381111561114d575f5ffd5b61115986828501610f68565b60e083015250949350505050565b5f60208284031215611177575f5ffd5b8151610d4781610c72565b6001600160a01b03841681526040602082018190528101829052818360608301375f818301606090810191909152601f909201601f1916010192915050565b5f602082840312156111d1575f5ffd5b8151610d4781610c1d565b634e487b7160e01b5f52603260045260245ffd5b602080825282518282018190525f918401906040840190835b81811015611227578351835260209384019390920191600101611209565b509095945050505050565b5f60208284031215611242575f5ffd5b81516001600160401b03811115611257575f5ffd5b8201601f81018413611267575f5ffd5b8051611275610eaf82610e70565b8082825260208201915060208360051b850101925086831115611296575f5ffd5b602084015b838110156113bc5780516001600160401b038111156112b8575f5ffd5b850160a0818a03601f190112156112cd575f5ffd5b6112d5610dad565b6020820151815260408201516001600160401b038111156112f4575f5ffd5b6113038b602083860101610dff565b60208301525060608201516040820152608082015161132181610c72565b606082015260a08201516001600160401b0381111561133e575f5ffd5b60208184010192505089601f830112611355575f5ffd5b8151611363610eaf82610e70565b8082825260208201915060208360051b86010192508c831115611384575f5ffd5b6020850194505b828510156113a657845182526020948501949091019061138b565b608084015250508452506020928301920161129b565b50969550505050505056fea26469706673582212202d6759bad135404a6aeb937cad6d5321510e2666340b06206a6174eadc4ab05764736f6c634300081c0033",
 }
 
 // AccessManagerABI is the input ABI used to generate the binding from.
