@@ -1364,53 +1364,6 @@ func (x *StreamFileDownloadChunkCreateResponse) GetBlocks() []*StreamFileDownloa
 	return nil
 }
 
-type StreamFileDownloadChunkCreateResponseV2 struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Blocks []*StreamFileDownloadChunkCreateResponseV2_BlockDownload `protobuf:"bytes,1,rep,name=blocks,proto3" json:"blocks,omitempty"`
-}
-
-func (x *StreamFileDownloadChunkCreateResponseV2) Reset() {
-	*x = StreamFileDownloadChunkCreateResponseV2{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[23]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
-}
-
-func (x *StreamFileDownloadChunkCreateResponseV2) String() string {
-	return protoimpl.X.MessageStringOf(x)
-}
-
-func (*StreamFileDownloadChunkCreateResponseV2) ProtoMessage() {}
-
-func (x *StreamFileDownloadChunkCreateResponseV2) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[23]
-	if protoimpl.UnsafeEnabled && x != nil {
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		if ms.LoadMessageInfo() == nil {
-			ms.StoreMessageInfo(mi)
-		}
-		return ms
-	}
-	return mi.MessageOf(x)
-}
-
-// Deprecated: Use StreamFileDownloadChunkCreateResponseV2.ProtoReflect.Descriptor instead.
-func (*StreamFileDownloadChunkCreateResponseV2) Descriptor() ([]byte, []int) {
-	return file_nodeapi_proto_rawDescGZIP(), []int{23}
-}
-
-func (x *StreamFileDownloadChunkCreateResponseV2) GetBlocks() []*StreamFileDownloadChunkCreateResponseV2_BlockDownload {
-	if x != nil {
-		return x.Blocks
-	}
-	return nil
-}
-
 type StreamFileDownloadBlockRequest struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
@@ -1426,7 +1379,7 @@ type StreamFileDownloadBlockRequest struct {
 func (x *StreamFileDownloadBlockRequest) Reset() {
 	*x = StreamFileDownloadBlockRequest{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[24]
+		mi := &file_nodeapi_proto_msgTypes[23]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -1439,7 +1392,7 @@ func (x *StreamFileDownloadBlockRequest) String() string {
 func (*StreamFileDownloadBlockRequest) ProtoMessage() {}
 
 func (x *StreamFileDownloadBlockRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[24]
+	mi := &file_nodeapi_proto_msgTypes[23]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -1452,7 +1405,7 @@ func (x *StreamFileDownloadBlockRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use StreamFileDownloadBlockRequest.ProtoReflect.Descriptor instead.
 func (*StreamFileDownloadBlockRequest) Descriptor() ([]byte, []int) {
-	return file_nodeapi_proto_rawDescGZIP(), []int{24}
+	return file_nodeapi_proto_rawDescGZIP(), []int{23}
 }
 
 func (x *StreamFileDownloadBlockRequest) GetStreamId() string {
@@ -1501,7 +1454,7 @@ type StreamFileListRequest struct {
 func (x *StreamFileListRequest) Reset() {
 	*x = StreamFileListRequest{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[25]
+		mi := &file_nodeapi_proto_msgTypes[24]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -1514,7 +1467,7 @@ func (x *StreamFileListRequest) String() string {
 func (*StreamFileListRequest) ProtoMessage() {}
 
 func (x *StreamFileListRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[25]
+	mi := &file_nodeapi_proto_msgTypes[24]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -1527,7 +1480,7 @@ func (x *StreamFileListRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use StreamFileListRequest.ProtoReflect.Descriptor instead.
 func (*StreamFileListRequest) Descriptor() ([]byte, []int) {
-	return file_nodeapi_proto_rawDescGZIP(), []int{25}
+	return file_nodeapi_proto_rawDescGZIP(), []int{24}
 }
 
 func (x *StreamFileListRequest) GetBucketName() string {
@@ -1556,7 +1509,7 @@ type File struct {
 func (x *File) Reset() {
 	*x = File{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[26]
+		mi := &file_nodeapi_proto_msgTypes[25]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -1569,7 +1522,7 @@ func (x *File) String() string {
 func (*File) ProtoMessage() {}
 
 func (x *File) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[26]
+	mi := &file_nodeapi_proto_msgTypes[25]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -1582,7 +1535,7 @@ func (x *File) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use File.ProtoReflect.Descriptor instead.
 func (*File) Descriptor() ([]byte, []int) {
-	return file_nodeapi_proto_rawDescGZIP(), []int{26}
+	return file_nodeapi_proto_rawDescGZIP(), []int{25}
 }
 
 func (x *File) GetStreamId() string {
@@ -1659,7 +1612,7 @@ type StreamFileListResponse struct {
 func (x *StreamFileListResponse) Reset() {
 	*x = StreamFileListResponse{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[27]
+		mi := &file_nodeapi_proto_msgTypes[26]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -1672,7 +1625,7 @@ func (x *StreamFileListResponse) String() string {
 func (*StreamFileListResponse) ProtoMessage() {}
 
 func (x *StreamFileListResponse) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[27]
+	mi := &file_nodeapi_proto_msgTypes[26]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -1685,7 +1638,7 @@ func (x *StreamFileListResponse) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use StreamFileListResponse.ProtoReflect.Descriptor instead.
 func (*StreamFileListResponse) Descriptor() ([]byte, []int) {
-	return file_nodeapi_proto_rawDescGZIP(), []int{27}
+	return file_nodeapi_proto_rawDescGZIP(), []int{26}
 }
 
 func (x *StreamFileListResponse) GetFiles() []*File {
@@ -1706,7 +1659,7 @@ type StreamFileListChunksRequest struct {
 func (x *StreamFileListChunksRequest) Reset() {
 	*x = StreamFileListChunksRequest{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[28]
+		mi := &file_nodeapi_proto_msgTypes[27]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -1719,7 +1672,7 @@ func (x *StreamFileListChunksRequest) String() string {
 func (*StreamFileListChunksRequest) ProtoMessage() {}
 
 func (x *StreamFileListChunksRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[28]
+	mi := &file_nodeapi_proto_msgTypes[27]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -1732,7 +1685,7 @@ func (x *StreamFileListChunksRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use StreamFileListChunksRequest.ProtoReflect.Descriptor instead.
 func (*StreamFileListChunksRequest) Descriptor() ([]byte, []int) {
-	return file_nodeapi_proto_rawDescGZIP(), []int{28}
+	return file_nodeapi_proto_rawDescGZIP(), []int{27}
 }
 
 func (x *StreamFileListChunksRequest) GetStreamId() string {
@@ -1754,7 +1707,7 @@ type StreamFileListChunksResponse struct {
 func (x *StreamFileListChunksResponse) Reset() {
 	*x = StreamFileListChunksResponse{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[29]
+		mi := &file_nodeapi_proto_msgTypes[28]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -1767,7 +1720,7 @@ func (x *StreamFileListChunksResponse) String() string {
 func (*StreamFileListChunksResponse) ProtoMessage() {}
 
 func (x *StreamFileListChunksResponse) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[29]
+	mi := &file_nodeapi_proto_msgTypes[28]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -1780,7 +1733,7 @@ func (x *StreamFileListChunksResponse) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use StreamFileListChunksResponse.ProtoReflect.Descriptor instead.
 func (*StreamFileListChunksResponse) Descriptor() ([]byte, []int) {
-	return file_nodeapi_proto_rawDescGZIP(), []int{29}
+	return file_nodeapi_proto_rawDescGZIP(), []int{28}
 }
 
 func (x *StreamFileListChunksResponse) GetStreamId() string {
@@ -1809,7 +1762,7 @@ type StreamFileViewRequest struct {
 func (x *StreamFileViewRequest) Reset() {
 	*x = StreamFileViewRequest{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[30]
+		mi := &file_nodeapi_proto_msgTypes[29]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -1822,7 +1775,7 @@ func (x *StreamFileViewRequest) String() string {
 func (*StreamFileViewRequest) ProtoMessage() {}
 
 func (x *StreamFileViewRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[30]
+	mi := &file_nodeapi_proto_msgTypes[29]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -1835,7 +1788,7 @@ func (x *StreamFileViewRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use StreamFileViewRequest.ProtoReflect.Descriptor instead.
 func (*StreamFileViewRequest) Descriptor() ([]byte, []int) {
-	return file_nodeapi_proto_rawDescGZIP(), []int{30}
+	return file_nodeapi_proto_rawDescGZIP(), []int{29}
 }
 
 func (x *StreamFileViewRequest) GetBucketName() string {
@@ -1872,7 +1825,7 @@ type StreamFileViewResponse struct {
 func (x *StreamFileViewResponse) Reset() {
 	*x = StreamFileViewResponse{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[31]
+		mi := &file_nodeapi_proto_msgTypes[30]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -1885,7 +1838,7 @@ func (x *StreamFileViewResponse) String() string {
 func (*StreamFileViewResponse) ProtoMessage() {}
 
 func (x *StreamFileViewResponse) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[31]
+	mi := &file_nodeapi_proto_msgTypes[30]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -1898,7 +1851,7 @@ func (x *StreamFileViewResponse) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use StreamFileViewResponse.ProtoReflect.Descriptor instead.
 func (*StreamFileViewResponse) Descriptor() ([]byte, []int) {
-	return file_nodeapi_proto_rawDescGZIP(), []int{31}
+	return file_nodeapi_proto_rawDescGZIP(), []int{30}
 }
 
 func (x *StreamFileViewResponse) GetBucketName() string {
@@ -1983,7 +1936,7 @@ type StreamFileDeleteRequest struct {
 func (x *StreamFileDeleteRequest) Reset() {
 	*x = StreamFileDeleteRequest{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[32]
+		mi := &file_nodeapi_proto_msgTypes[31]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -1996,7 +1949,7 @@ func (x *StreamFileDeleteRequest) String() string {
 func (*StreamFileDeleteRequest) ProtoMessage() {}
 
 func (x *StreamFileDeleteRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[32]
+	mi := &file_nodeapi_proto_msgTypes[31]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -2009,7 +1962,7 @@ func (x *StreamFileDeleteRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use StreamFileDeleteRequest.ProtoReflect.Descriptor instead.
 func (*StreamFileDeleteRequest) Descriptor() ([]byte, []int) {
-	return file_nodeapi_proto_rawDescGZIP(), []int{32}
+	return file_nodeapi_proto_rawDescGZIP(), []int{31}
 }
 
 func (x *StreamFileDeleteRequest) GetBucketName() string {
@@ -2035,7 +1988,7 @@ type StreamFileDeleteResponse struct {
 func (x *StreamFileDeleteResponse) Reset() {
 	*x = StreamFileDeleteResponse{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[33]
+		mi := &file_nodeapi_proto_msgTypes[32]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -2048,7 +2001,7 @@ func (x *StreamFileDeleteResponse) String() string {
 func (*StreamFileDeleteResponse) ProtoMessage() {}
 
 func (x *StreamFileDeleteResponse) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[33]
+	mi := &file_nodeapi_proto_msgTypes[32]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -2061,7 +2014,7 @@ func (x *StreamFileDeleteResponse) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use StreamFileDeleteResponse.ProtoReflect.Descriptor instead.
 func (*StreamFileDeleteResponse) Descriptor() ([]byte, []int) {
-	return file_nodeapi_proto_rawDescGZIP(), []int{33}
+	return file_nodeapi_proto_rawDescGZIP(), []int{32}
 }
 
 type StreamFileListVersionsRequest struct {
@@ -2076,7 +2029,7 @@ type StreamFileListVersionsRequest struct {
 func (x *StreamFileListVersionsRequest) Reset() {
 	*x = StreamFileListVersionsRequest{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[34]
+		mi := &file_nodeapi_proto_msgTypes[33]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -2089,7 +2042,7 @@ func (x *StreamFileListVersionsRequest) String() string {
 func (*StreamFileListVersionsRequest) ProtoMessage() {}
 
 func (x *StreamFileListVersionsRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[34]
+	mi := &file_nodeapi_proto_msgTypes[33]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -2102,7 +2055,7 @@ func (x *StreamFileListVersionsRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use StreamFileListVersionsRequest.ProtoReflect.Descriptor instead.
 func (*StreamFileListVersionsRequest) Descriptor() ([]byte, []int) {
-	return file_nodeapi_proto_rawDescGZIP(), []int{34}
+	return file_nodeapi_proto_rawDescGZIP(), []int{33}
 }
 
 func (x *StreamFileListVersionsRequest) GetBucketName() string {
@@ -2130,7 +2083,7 @@ type StreamFileListVersionsResponse struct {
 func (x *StreamFileListVersionsResponse) Reset() {
 	*x = StreamFileListVersionsResponse{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[35]
+		mi := &file_nodeapi_proto_msgTypes[34]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -2143,7 +2096,7 @@ func (x *StreamFileListVersionsResponse) String() string {
 func (*StreamFileListVersionsResponse) ProtoMessage() {}
 
 func (x *StreamFileListVersionsResponse) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[35]
+	mi := &file_nodeapi_proto_msgTypes[34]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -2156,7 +2109,7 @@ func (x *StreamFileListVersionsResponse) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use StreamFileListVersionsResponse.ProtoReflect.Descriptor instead.
 func (*StreamFileListVersionsResponse) Descriptor() ([]byte, []int) {
-	return file_nodeapi_proto_rawDescGZIP(), []int{35}
+	return file_nodeapi_proto_rawDescGZIP(), []int{34}
 }
 
 func (x *StreamFileListVersionsResponse) GetVersions() []*File {
@@ -2178,7 +2131,7 @@ type BucketListResponse_Bucket struct {
 func (x *BucketListResponse_Bucket) Reset() {
 	*x = BucketListResponse_Bucket{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[36]
+		mi := &file_nodeapi_proto_msgTypes[35]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -2191,7 +2144,7 @@ func (x *BucketListResponse_Bucket) String() string {
 func (*BucketListResponse_Bucket) ProtoMessage() {}
 
 func (x *BucketListResponse_Bucket) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[36]
+	mi := &file_nodeapi_proto_msgTypes[35]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -2233,7 +2186,7 @@ type Chunk_Block struct {
 func (x *Chunk_Block) Reset() {
 	*x = Chunk_Block{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[37]
+		mi := &file_nodeapi_proto_msgTypes[36]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -2246,7 +2199,7 @@ func (x *Chunk_Block) String() string {
 func (*Chunk_Block) ProtoMessage() {}
 
 func (x *Chunk_Block) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[37]
+	mi := &file_nodeapi_proto_msgTypes[36]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -2290,7 +2243,7 @@ type StreamFileUploadChunkCreateResponse_BlockUpload struct {
 func (x *StreamFileUploadChunkCreateResponse_BlockUpload) Reset() {
 	*x = StreamFileUploadChunkCreateResponse_BlockUpload{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[38]
+		mi := &file_nodeapi_proto_msgTypes[37]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -2303,7 +2256,7 @@ func (x *StreamFileUploadChunkCreateResponse_BlockUpload) String() string {
 func (*StreamFileUploadChunkCreateResponse_BlockUpload) ProtoMessage() {}
 
 func (x *StreamFileUploadChunkCreateResponse_BlockUpload) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[38]
+	mi := &file_nodeapi_proto_msgTypes[37]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -2360,7 +2313,7 @@ type StreamFileDownloadCreateResponse_Chunk struct {
 func (x *StreamFileDownloadCreateResponse_Chunk) Reset() {
 	*x = StreamFileDownloadCreateResponse_Chunk{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[39]
+		mi := &file_nodeapi_proto_msgTypes[38]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -2373,7 +2326,7 @@ func (x *StreamFileDownloadCreateResponse_Chunk) String() string {
 func (*StreamFileDownloadCreateResponse_Chunk) ProtoMessage() {}
 
 func (x *StreamFileDownloadCreateResponse_Chunk) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[39]
+	mi := &file_nodeapi_proto_msgTypes[38]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -2425,7 +2378,7 @@ type StreamFileDownloadChunkCreateResponse_BlockDownload struct {
 func (x *StreamFileDownloadChunkCreateResponse_BlockDownload) Reset() {
 	*x = StreamFileDownloadChunkCreateResponse_BlockDownload{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[40]
+		mi := &file_nodeapi_proto_msgTypes[39]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -2438,7 +2391,7 @@ func (x *StreamFileDownloadChunkCreateResponse_BlockDownload) String() string {
 func (*StreamFileDownloadChunkCreateResponse_BlockDownload) ProtoMessage() {}
 
 func (x *StreamFileDownloadChunkCreateResponse_BlockDownload) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[40]
+	mi := &file_nodeapi_proto_msgTypes[39]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -2485,179 +2438,6 @@ func (x *StreamFileDownloadChunkCreateResponse_BlockDownload) GetNodeId() string
 func (x *StreamFileDownloadChunkCreateResponse_BlockDownload) GetPermit() string {
 	if x != nil {
 		return x.Permit
-	}
-	return ""
-}
-
-type StreamFileDownloadChunkCreateResponseV2_BlockDownload struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Cid      string                                                          `protobuf:"bytes,1,opt,name=cid,proto3" json:"cid,omitempty"`
-	Size     int64                                                           `protobuf:"varint,2,opt,name=size,proto3" json:"size,omitempty"`
-	Akave    *StreamFileDownloadChunkCreateResponseV2_BlockDownload_Akave    `protobuf:"bytes,3,opt,name=akave,proto3" json:"akave,omitempty"`
-	Filecoin *StreamFileDownloadChunkCreateResponseV2_BlockDownload_Filecoin `protobuf:"bytes,4,opt,name=filecoin,proto3" json:"filecoin,omitempty"`
-}
-
-func (x *StreamFileDownloadChunkCreateResponseV2_BlockDownload) Reset() {
-	*x = StreamFileDownloadChunkCreateResponseV2_BlockDownload{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[41]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
-}
-
-func (x *StreamFileDownloadChunkCreateResponseV2_BlockDownload) String() string {
-	return protoimpl.X.MessageStringOf(x)
-}
-
-func (*StreamFileDownloadChunkCreateResponseV2_BlockDownload) ProtoMessage() {}
-
-func (x *StreamFileDownloadChunkCreateResponseV2_BlockDownload) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[41]
-	if protoimpl.UnsafeEnabled && x != nil {
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		if ms.LoadMessageInfo() == nil {
-			ms.StoreMessageInfo(mi)
-		}
-		return ms
-	}
-	return mi.MessageOf(x)
-}
-
-// Deprecated: Use StreamFileDownloadChunkCreateResponseV2_BlockDownload.ProtoReflect.Descriptor instead.
-func (*StreamFileDownloadChunkCreateResponseV2_BlockDownload) Descriptor() ([]byte, []int) {
-	return file_nodeapi_proto_rawDescGZIP(), []int{23, 0}
-}
-
-func (x *StreamFileDownloadChunkCreateResponseV2_BlockDownload) GetCid() string {
-	if x != nil {
-		return x.Cid
-	}
-	return ""
-}
-
-func (x *StreamFileDownloadChunkCreateResponseV2_BlockDownload) GetSize() int64 {
-	if x != nil {
-		return x.Size
-	}
-	return 0
-}
-
-func (x *StreamFileDownloadChunkCreateResponseV2_BlockDownload) GetAkave() *StreamFileDownloadChunkCreateResponseV2_BlockDownload_Akave {
-	if x != nil {
-		return x.Akave
-	}
-	return nil
-}
-
-func (x *StreamFileDownloadChunkCreateResponseV2_BlockDownload) GetFilecoin() *StreamFileDownloadChunkCreateResponseV2_BlockDownload_Filecoin {
-	if x != nil {
-		return x.Filecoin
-	}
-	return nil
-}
-
-type StreamFileDownloadChunkCreateResponseV2_BlockDownload_Akave struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	NodeId      string `protobuf:"bytes,1,opt,name=node_id,json=nodeId,proto3" json:"node_id,omitempty"`
-	NodeAddress string `protobuf:"bytes,2,opt,name=node_address,json=nodeAddress,proto3" json:"node_address,omitempty"`
-}
-
-func (x *StreamFileDownloadChunkCreateResponseV2_BlockDownload_Akave) Reset() {
-	*x = StreamFileDownloadChunkCreateResponseV2_BlockDownload_Akave{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[42]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
-}
-
-func (x *StreamFileDownloadChunkCreateResponseV2_BlockDownload_Akave) String() string {
-	return protoimpl.X.MessageStringOf(x)
-}
-
-func (*StreamFileDownloadChunkCreateResponseV2_BlockDownload_Akave) ProtoMessage() {}
-
-func (x *StreamFileDownloadChunkCreateResponseV2_BlockDownload_Akave) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[42]
-	if protoimpl.UnsafeEnabled && x != nil {
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		if ms.LoadMessageInfo() == nil {
-			ms.StoreMessageInfo(mi)
-		}
-		return ms
-	}
-	return mi.MessageOf(x)
-}
-
-// Deprecated: Use StreamFileDownloadChunkCreateResponseV2_BlockDownload_Akave.ProtoReflect.Descriptor instead.
-func (*StreamFileDownloadChunkCreateResponseV2_BlockDownload_Akave) Descriptor() ([]byte, []int) {
-	return file_nodeapi_proto_rawDescGZIP(), []int{23, 0, 0}
-}
-
-func (x *StreamFileDownloadChunkCreateResponseV2_BlockDownload_Akave) GetNodeId() string {
-	if x != nil {
-		return x.NodeId
-	}
-	return ""
-}
-
-func (x *StreamFileDownloadChunkCreateResponseV2_BlockDownload_Akave) GetNodeAddress() string {
-	if x != nil {
-		return x.NodeAddress
-	}
-	return ""
-}
-
-type StreamFileDownloadChunkCreateResponseV2_BlockDownload_Filecoin struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	SpAddress string `protobuf:"bytes,1,opt,name=sp_address,json=spAddress,proto3" json:"sp_address,omitempty"`
-}
-
-func (x *StreamFileDownloadChunkCreateResponseV2_BlockDownload_Filecoin) Reset() {
-	*x = StreamFileDownloadChunkCreateResponseV2_BlockDownload_Filecoin{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_nodeapi_proto_msgTypes[43]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
-}
-
-func (x *StreamFileDownloadChunkCreateResponseV2_BlockDownload_Filecoin) String() string {
-	return protoimpl.X.MessageStringOf(x)
-}
-
-func (*StreamFileDownloadChunkCreateResponseV2_BlockDownload_Filecoin) ProtoMessage() {}
-
-func (x *StreamFileDownloadChunkCreateResponseV2_BlockDownload_Filecoin) ProtoReflect() protoreflect.Message {
-	mi := &file_nodeapi_proto_msgTypes[43]
-	if protoimpl.UnsafeEnabled && x != nil {
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		if ms.LoadMessageInfo() == nil {
-			ms.StoreMessageInfo(mi)
-		}
-		return ms
-	}
-	return mi.MessageOf(x)
-}
-
-// Deprecated: Use StreamFileDownloadChunkCreateResponseV2_BlockDownload_Filecoin.ProtoReflect.Descriptor instead.
-func (*StreamFileDownloadChunkCreateResponseV2_BlockDownload_Filecoin) Descriptor() ([]byte, []int) {
-	return file_nodeapi_proto_rawDescGZIP(), []int{23, 0, 1}
-}
-
-func (x *StreamFileDownloadChunkCreateResponseV2_BlockDownload_Filecoin) GetSpAddress() string {
-	if x != nil {
-		return x.SpAddress
 	}
 	return ""
 }
@@ -2859,253 +2639,214 @@ var file_nodeapi_proto_rawDesc = []byte{
 	0x0b, 0x6e, 0x6f, 0x64, 0x65, 0x41, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x12, 0x17, 0x0a, 0x07,
 	0x6e, 0x6f, 0x64, 0x65, 0x5f, 0x69, 0x64, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x6e,
 	0x6f, 0x64, 0x65, 0x49, 0x64, 0x12, 0x16, 0x0a, 0x06, 0x70, 0x65, 0x72, 0x6d, 0x69, 0x74, 0x18,
-	0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x70, 0x65, 0x72, 0x6d, 0x69, 0x74, 0x22, 0xea, 0x03,
-	0x0a, 0x27, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e,
-	0x6c, 0x6f, 0x61, 0x64, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x52,
-	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x56, 0x32, 0x12, 0x56, 0x0a, 0x06, 0x62, 0x6c, 0x6f,
-	0x63, 0x6b, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x3e, 0x2e, 0x6e, 0x6f, 0x64, 0x65,
-	0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f,
-	0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x43, 0x72, 0x65, 0x61, 0x74,
-	0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x56, 0x32, 0x2e, 0x42, 0x6c, 0x6f, 0x63,
-	0x6b, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x52, 0x06, 0x62, 0x6c, 0x6f, 0x63, 0x6b,
-	0x73, 0x1a, 0xe6, 0x02, 0x0a, 0x0d, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x44, 0x6f, 0x77, 0x6e, 0x6c,
-	0x6f, 0x61, 0x64, 0x12, 0x10, 0x0a, 0x03, 0x63, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09,
-	0x52, 0x03, 0x63, 0x69, 0x64, 0x12, 0x12, 0x0a, 0x04, 0x73, 0x69, 0x7a, 0x65, 0x18, 0x02, 0x20,
-	0x01, 0x28, 0x03, 0x52, 0x04, 0x73, 0x69, 0x7a, 0x65, 0x12, 0x5a, 0x0a, 0x05, 0x61, 0x6b, 0x61,
-	0x76, 0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x44, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61,
-	0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77,
-	0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65,
-	0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x56, 0x32, 0x2e, 0x42, 0x6c, 0x6f, 0x63, 0x6b,
-	0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x2e, 0x41, 0x6b, 0x61, 0x76, 0x65, 0x52, 0x05,
-	0x61, 0x6b, 0x61, 0x76, 0x65, 0x12, 0x63, 0x0a, 0x08, 0x66, 0x69, 0x6c, 0x65, 0x63, 0x6f, 0x69,
-	0x6e, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x47, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70,
-	0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e,
-	0x6c, 0x6f, 0x61, 0x64, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x52,
-	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x56, 0x32, 0x2e, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x44,
-	0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x2e, 0x46, 0x69, 0x6c, 0x65, 0x63, 0x6f, 0x69, 0x6e,
-	0x52, 0x08, 0x66, 0x69, 0x6c, 0x65, 0x63, 0x6f, 0x69, 0x6e, 0x1a, 0x43, 0x0a, 0x05, 0x41, 0x6b,
-	0x61, 0x76, 0x65, 0x12, 0x17, 0x0a, 0x07, 0x6e, 0x6f, 0x64, 0x65, 0x5f, 0x69, 0x64, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x6e, 0x6f, 0x64, 0x65, 0x49, 0x64, 0x12, 0x21, 0x0a, 0x0c,
-	0x6e, 0x6f, 0x64, 0x65, 0x5f, 0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x18, 0x02, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x0b, 0x6e, 0x6f, 0x64, 0x65, 0x41, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x1a,
-	0x29, 0x0a, 0x08, 0x46, 0x69, 0x6c, 0x65, 0x63, 0x6f, 0x69, 0x6e, 0x12, 0x1d, 0x0a, 0x0a, 0x73,
-	0x70, 0x5f, 0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52,
-	0x09, 0x73, 0x70, 0x41, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x22, 0xb9, 0x01, 0x0a, 0x1e, 0x53,
-	0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61,
-	0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1b, 0x0a,
-	0x09, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x5f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09,
-	0x52, 0x08, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x49, 0x64, 0x12, 0x1b, 0x0a, 0x09, 0x63, 0x68,
-	0x75, 0x6e, 0x6b, 0x5f, 0x63, 0x69, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x63,
-	0x68, 0x75, 0x6e, 0x6b, 0x43, 0x69, 0x64, 0x12, 0x1f, 0x0a, 0x0b, 0x63, 0x68, 0x75, 0x6e, 0x6b,
-	0x5f, 0x69, 0x6e, 0x64, 0x65, 0x78, 0x18, 0x03, 0x20, 0x01, 0x28, 0x03, 0x52, 0x0a, 0x63, 0x68,
-	0x75, 0x6e, 0x6b, 0x49, 0x6e, 0x64, 0x65, 0x78, 0x12, 0x1b, 0x0a, 0x09, 0x62, 0x6c, 0x6f, 0x63,
-	0x6b, 0x5f, 0x63, 0x69, 0x64, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x62, 0x6c, 0x6f,
-	0x63, 0x6b, 0x43, 0x69, 0x64, 0x12, 0x1f, 0x0a, 0x0b, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x69,
-	0x6e, 0x64, 0x65, 0x78, 0x18, 0x05, 0x20, 0x01, 0x28, 0x03, 0x52, 0x0a, 0x62, 0x6c, 0x6f, 0x63,
-	0x6b, 0x49, 0x6e, 0x64, 0x65, 0x78, 0x22, 0x38, 0x0a, 0x15, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d,
-	0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12,
+	0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x70, 0x65, 0x72, 0x6d, 0x69, 0x74, 0x22, 0xb9, 0x01,
+	0x0a, 0x1e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e,
+	0x6c, 0x6f, 0x61, 0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
+	0x12, 0x1b, 0x0a, 0x09, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x5f, 0x69, 0x64, 0x18, 0x01, 0x20,
+	0x01, 0x28, 0x09, 0x52, 0x08, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x49, 0x64, 0x12, 0x1b, 0x0a,
+	0x09, 0x63, 0x68, 0x75, 0x6e, 0x6b, 0x5f, 0x63, 0x69, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09,
+	0x52, 0x08, 0x63, 0x68, 0x75, 0x6e, 0x6b, 0x43, 0x69, 0x64, 0x12, 0x1f, 0x0a, 0x0b, 0x63, 0x68,
+	0x75, 0x6e, 0x6b, 0x5f, 0x69, 0x6e, 0x64, 0x65, 0x78, 0x18, 0x03, 0x20, 0x01, 0x28, 0x03, 0x52,
+	0x0a, 0x63, 0x68, 0x75, 0x6e, 0x6b, 0x49, 0x6e, 0x64, 0x65, 0x78, 0x12, 0x1b, 0x0a, 0x09, 0x62,
+	0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x63, 0x69, 0x64, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08,
+	0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x43, 0x69, 0x64, 0x12, 0x1f, 0x0a, 0x0b, 0x62, 0x6c, 0x6f, 0x63,
+	0x6b, 0x5f, 0x69, 0x6e, 0x64, 0x65, 0x78, 0x18, 0x05, 0x20, 0x01, 0x28, 0x03, 0x52, 0x0a, 0x62,
+	0x6c, 0x6f, 0x63, 0x6b, 0x49, 0x6e, 0x64, 0x65, 0x78, 0x22, 0x38, 0x0a, 0x15, 0x53, 0x74, 0x72,
+	0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65,
+	0x73, 0x74, 0x12, 0x1f, 0x0a, 0x0b, 0x62, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x5f, 0x6e, 0x61, 0x6d,
+	0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x62, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x4e,
+	0x61, 0x6d, 0x65, 0x22, 0xc5, 0x02, 0x0a, 0x04, 0x46, 0x69, 0x6c, 0x65, 0x12, 0x1b, 0x0a, 0x09,
+	0x73, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x5f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52,
+	0x08, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x49, 0x64, 0x12, 0x19, 0x0a, 0x08, 0x72, 0x6f, 0x6f,
+	0x74, 0x5f, 0x63, 0x69, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x72, 0x6f, 0x6f,
+	0x74, 0x43, 0x69, 0x64, 0x12, 0x12, 0x0a, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x03, 0x20, 0x01,
+	0x28, 0x09, 0x52, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x12, 0x21, 0x0a, 0x0c, 0x65, 0x6e, 0x63, 0x6f,
+	0x64, 0x65, 0x64, 0x5f, 0x73, 0x69, 0x7a, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28, 0x03, 0x52, 0x0b,
+	0x65, 0x6e, 0x63, 0x6f, 0x64, 0x65, 0x64, 0x53, 0x69, 0x7a, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x73,
+	0x69, 0x7a, 0x65, 0x18, 0x05, 0x20, 0x01, 0x28, 0x03, 0x52, 0x04, 0x73, 0x69, 0x7a, 0x65, 0x12,
+	0x1f, 0x0a, 0x0b, 0x64, 0x61, 0x74, 0x61, 0x5f, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x18, 0x06,
+	0x20, 0x01, 0x28, 0x03, 0x52, 0x0a, 0x64, 0x61, 0x74, 0x61, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73,
+	0x12, 0x21, 0x0a, 0x0c, 0x74, 0x6f, 0x74, 0x61, 0x6c, 0x5f, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73,
+	0x18, 0x07, 0x20, 0x01, 0x28, 0x03, 0x52, 0x0b, 0x74, 0x6f, 0x74, 0x61, 0x6c, 0x42, 0x6c, 0x6f,
+	0x63, 0x6b, 0x73, 0x12, 0x39, 0x0a, 0x0a, 0x63, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x5f, 0x61,
+	0x74, 0x18, 0x08, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65,
+	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74,
+	0x61, 0x6d, 0x70, 0x52, 0x09, 0x63, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x41, 0x74, 0x12, 0x3b,
+	0x0a, 0x0b, 0x63, 0x6f, 0x6d, 0x6d, 0x69, 0x74, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x09, 0x20,
+	0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f,
+	0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52,
+	0x0a, 0x63, 0x6f, 0x6d, 0x6d, 0x69, 0x74, 0x65, 0x64, 0x41, 0x74, 0x22, 0x3d, 0x0a, 0x16, 0x53,
+	0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x73,
+	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x23, 0x0a, 0x05, 0x66, 0x69, 0x6c, 0x65, 0x73, 0x18, 0x01,
+	0x20, 0x03, 0x28, 0x0b, 0x32, 0x0d, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x46,
+	0x69, 0x6c, 0x65, 0x52, 0x05, 0x66, 0x69, 0x6c, 0x65, 0x73, 0x22, 0x3a, 0x0a, 0x1b, 0x53, 0x74,
+	0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x43, 0x68, 0x75, 0x6e,
+	0x6b, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1b, 0x0a, 0x09, 0x73, 0x74, 0x72,
+	0x65, 0x61, 0x6d, 0x5f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x73, 0x74,
+	0x72, 0x65, 0x61, 0x6d, 0x49, 0x64, 0x22, 0x4f, 0x0a, 0x1c, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d,
+	0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x73, 0x52, 0x65,
+	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x1b, 0x0a, 0x09, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d,
+	0x5f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x73, 0x74, 0x72, 0x65, 0x61,
+	0x6d, 0x49, 0x64, 0x12, 0x12, 0x0a, 0x04, 0x63, 0x69, 0x64, 0x73, 0x18, 0x02, 0x20, 0x03, 0x28,
+	0x09, 0x52, 0x04, 0x63, 0x69, 0x64, 0x73, 0x22, 0x55, 0x0a, 0x15, 0x53, 0x74, 0x72, 0x65, 0x61,
+	0x6d, 0x46, 0x69, 0x6c, 0x65, 0x56, 0x69, 0x65, 0x77, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
+	0x12, 0x1f, 0x0a, 0x0b, 0x62, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18,
+	0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x62, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x4e, 0x61, 0x6d,
+	0x65, 0x12, 0x1b, 0x0a, 0x09, 0x66, 0x69, 0x6c, 0x65, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x02,
+	0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x66, 0x69, 0x6c, 0x65, 0x4e, 0x61, 0x6d, 0x65, 0x22, 0x83,
+	0x03, 0x0a, 0x16, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x56, 0x69, 0x65,
+	0x77, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x1f, 0x0a, 0x0b, 0x62, 0x75, 0x63,
+	0x6b, 0x65, 0x74, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a,
+	0x62, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1b, 0x0a, 0x09, 0x66, 0x69,
+	0x6c, 0x65, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x66,
+	0x69, 0x6c, 0x65, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1b, 0x0a, 0x09, 0x73, 0x74, 0x72, 0x65, 0x61,
+	0x6d, 0x5f, 0x69, 0x64, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x73, 0x74, 0x72, 0x65,
+	0x61, 0x6d, 0x49, 0x64, 0x12, 0x19, 0x0a, 0x08, 0x72, 0x6f, 0x6f, 0x74, 0x5f, 0x63, 0x69, 0x64,
+	0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x72, 0x6f, 0x6f, 0x74, 0x43, 0x69, 0x64, 0x12,
+	0x21, 0x0a, 0x0c, 0x65, 0x6e, 0x63, 0x6f, 0x64, 0x65, 0x64, 0x5f, 0x73, 0x69, 0x7a, 0x65, 0x18,
+	0x05, 0x20, 0x01, 0x28, 0x03, 0x52, 0x0b, 0x65, 0x6e, 0x63, 0x6f, 0x64, 0x65, 0x64, 0x53, 0x69,
+	0x7a, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x73, 0x69, 0x7a, 0x65, 0x18, 0x06, 0x20, 0x01, 0x28, 0x03,
+	0x52, 0x04, 0x73, 0x69, 0x7a, 0x65, 0x12, 0x1f, 0x0a, 0x0b, 0x64, 0x61, 0x74, 0x61, 0x5f, 0x62,
+	0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x18, 0x07, 0x20, 0x01, 0x28, 0x03, 0x52, 0x0a, 0x64, 0x61, 0x74,
+	0x61, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x12, 0x21, 0x0a, 0x0c, 0x74, 0x6f, 0x74, 0x61, 0x6c,
+	0x5f, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x18, 0x08, 0x20, 0x01, 0x28, 0x03, 0x52, 0x0b, 0x74,
+	0x6f, 0x74, 0x61, 0x6c, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x12, 0x39, 0x0a, 0x0a, 0x63, 0x72,
+	0x65, 0x61, 0x74, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x09, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a,
+	0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66,
+	0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x09, 0x63, 0x72, 0x65, 0x61,
+	0x74, 0x65, 0x64, 0x41, 0x74, 0x12, 0x3d, 0x0a, 0x0c, 0x63, 0x6f, 0x6d, 0x6d, 0x69, 0x74, 0x74,
+	0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x0a, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f,
+	0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69,
+	0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x0b, 0x63, 0x6f, 0x6d, 0x6d, 0x69, 0x74, 0x74,
+	0x65, 0x64, 0x41, 0x74, 0x22, 0x57, 0x0a, 0x17, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69,
+	0x6c, 0x65, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12,
 	0x1f, 0x0a, 0x0b, 0x62, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01,
 	0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x62, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x4e, 0x61, 0x6d, 0x65,
-	0x22, 0xc5, 0x02, 0x0a, 0x04, 0x46, 0x69, 0x6c, 0x65, 0x12, 0x1b, 0x0a, 0x09, 0x73, 0x74, 0x72,
-	0x65, 0x61, 0x6d, 0x5f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x73, 0x74,
-	0x72, 0x65, 0x61, 0x6d, 0x49, 0x64, 0x12, 0x19, 0x0a, 0x08, 0x72, 0x6f, 0x6f, 0x74, 0x5f, 0x63,
-	0x69, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x72, 0x6f, 0x6f, 0x74, 0x43, 0x69,
-	0x64, 0x12, 0x12, 0x0a, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52,
-	0x04, 0x6e, 0x61, 0x6d, 0x65, 0x12, 0x21, 0x0a, 0x0c, 0x65, 0x6e, 0x63, 0x6f, 0x64, 0x65, 0x64,
-	0x5f, 0x73, 0x69, 0x7a, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28, 0x03, 0x52, 0x0b, 0x65, 0x6e, 0x63,
-	0x6f, 0x64, 0x65, 0x64, 0x53, 0x69, 0x7a, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x73, 0x69, 0x7a, 0x65,
-	0x18, 0x05, 0x20, 0x01, 0x28, 0x03, 0x52, 0x04, 0x73, 0x69, 0x7a, 0x65, 0x12, 0x1f, 0x0a, 0x0b,
-	0x64, 0x61, 0x74, 0x61, 0x5f, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x18, 0x06, 0x20, 0x01, 0x28,
-	0x03, 0x52, 0x0a, 0x64, 0x61, 0x74, 0x61, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x12, 0x21, 0x0a,
-	0x0c, 0x74, 0x6f, 0x74, 0x61, 0x6c, 0x5f, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x18, 0x07, 0x20,
-	0x01, 0x28, 0x03, 0x52, 0x0b, 0x74, 0x6f, 0x74, 0x61, 0x6c, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73,
-	0x12, 0x39, 0x0a, 0x0a, 0x63, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x08,
-	0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70,
-	0x52, 0x09, 0x63, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x41, 0x74, 0x12, 0x3b, 0x0a, 0x0b, 0x63,
-	0x6f, 0x6d, 0x6d, 0x69, 0x74, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x09, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62,
-	0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x0a, 0x63, 0x6f,
-	0x6d, 0x6d, 0x69, 0x74, 0x65, 0x64, 0x41, 0x74, 0x22, 0x3d, 0x0a, 0x16, 0x53, 0x74, 0x72, 0x65,
-	0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e,
-	0x73, 0x65, 0x12, 0x23, 0x0a, 0x05, 0x66, 0x69, 0x6c, 0x65, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28,
-	0x0b, 0x32, 0x0d, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x46, 0x69, 0x6c, 0x65,
-	0x52, 0x05, 0x66, 0x69, 0x6c, 0x65, 0x73, 0x22, 0x3a, 0x0a, 0x1b, 0x53, 0x74, 0x72, 0x65, 0x61,
-	0x6d, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x73, 0x52,
-	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1b, 0x0a, 0x09, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d,
-	0x5f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x73, 0x74, 0x72, 0x65, 0x61,
-	0x6d, 0x49, 0x64, 0x22, 0x4f, 0x0a, 0x1c, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c,
-	0x65, 0x4c, 0x69, 0x73, 0x74, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f,
-	0x6e, 0x73, 0x65, 0x12, 0x1b, 0x0a, 0x09, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x5f, 0x69, 0x64,
-	0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x49, 0x64,
-	0x12, 0x12, 0x0a, 0x04, 0x63, 0x69, 0x64, 0x73, 0x18, 0x02, 0x20, 0x03, 0x28, 0x09, 0x52, 0x04,
-	0x63, 0x69, 0x64, 0x73, 0x22, 0x55, 0x0a, 0x15, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69,
-	0x6c, 0x65, 0x56, 0x69, 0x65, 0x77, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1f, 0x0a,
-	0x0b, 0x62, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x0a, 0x62, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1b,
-	0x0a, 0x09, 0x66, 0x69, 0x6c, 0x65, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28,
-	0x09, 0x52, 0x08, 0x66, 0x69, 0x6c, 0x65, 0x4e, 0x61, 0x6d, 0x65, 0x22, 0x83, 0x03, 0x0a, 0x16,
-	0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x56, 0x69, 0x65, 0x77, 0x52, 0x65,
-	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x1f, 0x0a, 0x0b, 0x62, 0x75, 0x63, 0x6b, 0x65, 0x74,
-	0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x62, 0x75, 0x63,
-	0x6b, 0x65, 0x74, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1b, 0x0a, 0x09, 0x66, 0x69, 0x6c, 0x65, 0x5f,
-	0x6e, 0x61, 0x6d, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x66, 0x69, 0x6c, 0x65,
-	0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1b, 0x0a, 0x09, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x5f, 0x69,
-	0x64, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x49,
-	0x64, 0x12, 0x19, 0x0a, 0x08, 0x72, 0x6f, 0x6f, 0x74, 0x5f, 0x63, 0x69, 0x64, 0x18, 0x04, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x07, 0x72, 0x6f, 0x6f, 0x74, 0x43, 0x69, 0x64, 0x12, 0x21, 0x0a, 0x0c,
-	0x65, 0x6e, 0x63, 0x6f, 0x64, 0x65, 0x64, 0x5f, 0x73, 0x69, 0x7a, 0x65, 0x18, 0x05, 0x20, 0x01,
-	0x28, 0x03, 0x52, 0x0b, 0x65, 0x6e, 0x63, 0x6f, 0x64, 0x65, 0x64, 0x53, 0x69, 0x7a, 0x65, 0x12,
-	0x12, 0x0a, 0x04, 0x73, 0x69, 0x7a, 0x65, 0x18, 0x06, 0x20, 0x01, 0x28, 0x03, 0x52, 0x04, 0x73,
-	0x69, 0x7a, 0x65, 0x12, 0x1f, 0x0a, 0x0b, 0x64, 0x61, 0x74, 0x61, 0x5f, 0x62, 0x6c, 0x6f, 0x63,
-	0x6b, 0x73, 0x18, 0x07, 0x20, 0x01, 0x28, 0x03, 0x52, 0x0a, 0x64, 0x61, 0x74, 0x61, 0x42, 0x6c,
-	0x6f, 0x63, 0x6b, 0x73, 0x12, 0x21, 0x0a, 0x0c, 0x74, 0x6f, 0x74, 0x61, 0x6c, 0x5f, 0x62, 0x6c,
-	0x6f, 0x63, 0x6b, 0x73, 0x18, 0x08, 0x20, 0x01, 0x28, 0x03, 0x52, 0x0b, 0x74, 0x6f, 0x74, 0x61,
-	0x6c, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x12, 0x39, 0x0a, 0x0a, 0x63, 0x72, 0x65, 0x61, 0x74,
-	0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x09, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f,
-	0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69,
-	0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x09, 0x63, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64,
-	0x41, 0x74, 0x12, 0x3d, 0x0a, 0x0c, 0x63, 0x6f, 0x6d, 0x6d, 0x69, 0x74, 0x74, 0x65, 0x64, 0x5f,
-	0x61, 0x74, 0x18, 0x0a, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c,
-	0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73,
-	0x74, 0x61, 0x6d, 0x70, 0x52, 0x0b, 0x63, 0x6f, 0x6d, 0x6d, 0x69, 0x74, 0x74, 0x65, 0x64, 0x41,
-	0x74, 0x22, 0x57, 0x0a, 0x17, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44,
-	0x65, 0x6c, 0x65, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1f, 0x0a, 0x0b,
-	0x62, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28,
-	0x09, 0x52, 0x0a, 0x62, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1b, 0x0a,
-	0x09, 0x66, 0x69, 0x6c, 0x65, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09,
-	0x52, 0x08, 0x66, 0x69, 0x6c, 0x65, 0x4e, 0x61, 0x6d, 0x65, 0x22, 0x1a, 0x0a, 0x18, 0x53, 0x74,
-	0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x52, 0x65,
-	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x5d, 0x0a, 0x1d, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d,
-	0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x73,
-	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1f, 0x0a, 0x0b, 0x62, 0x75, 0x63, 0x6b, 0x65,
-	0x74, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x62, 0x75,
-	0x63, 0x6b, 0x65, 0x74, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1b, 0x0a, 0x09, 0x66, 0x69, 0x6c, 0x65,
-	0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x66, 0x69, 0x6c,
-	0x65, 0x4e, 0x61, 0x6d, 0x65, 0x22, 0x4b, 0x0a, 0x1e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46,
-	0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x73, 0x52,
-	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x29, 0x0a, 0x08, 0x76, 0x65, 0x72, 0x73, 0x69,
-	0x6f, 0x6e, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x0d, 0x2e, 0x6e, 0x6f, 0x64, 0x65,
-	0x61, 0x70, 0x69, 0x2e, 0x46, 0x69, 0x6c, 0x65, 0x52, 0x08, 0x76, 0x65, 0x72, 0x73, 0x69, 0x6f,
-	0x6e, 0x73, 0x32, 0xb9, 0x02, 0x0a, 0x07, 0x4e, 0x6f, 0x64, 0x65, 0x41, 0x50, 0x49, 0x12, 0x4d,
-	0x0a, 0x0c, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x12, 0x1c,
-	0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x43,
-	0x72, 0x65, 0x61, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1d, 0x2e, 0x6e,
-	0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x43, 0x72, 0x65,
-	0x61, 0x74, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x47, 0x0a,
-	0x0a, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x56, 0x69, 0x65, 0x77, 0x12, 0x1a, 0x2e, 0x6e, 0x6f,
-	0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x56, 0x69, 0x65, 0x77,
-	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1b, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70,
-	0x69, 0x2e, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x56, 0x69, 0x65, 0x77, 0x52, 0x65, 0x73, 0x70,
-	0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x47, 0x0a, 0x0a, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74,
-	0x4c, 0x69, 0x73, 0x74, 0x12, 0x1a, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x42,
-	0x75, 0x63, 0x6b, 0x65, 0x74, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
-	0x1a, 0x1b, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x42, 0x75, 0x63, 0x6b, 0x65,
-	0x74, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12,
-	0x4d, 0x0a, 0x0c, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x12,
-	0x1c, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74,
-	0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1d, 0x2e,
-	0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x44, 0x65,
-	0x6c, 0x65, 0x74, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x32, 0x83,
-	0x0c, 0x0a, 0x09, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x41, 0x50, 0x49, 0x12, 0x65, 0x0a, 0x10,
-	0x46, 0x69, 0x6c, 0x65, 0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65,
-	0x12, 0x26, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61,
-	0x6d, 0x46, 0x69, 0x6c, 0x65, 0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x72, 0x65, 0x61, 0x74,
-	0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x27, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61,
-	0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x55, 0x70, 0x6c,
-	0x6f, 0x61, 0x64, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73,
-	0x65, 0x22, 0x00, 0x12, 0x74, 0x0a, 0x15, 0x46, 0x69, 0x6c, 0x65, 0x55, 0x70, 0x6c, 0x6f, 0x61,
-	0x64, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x12, 0x2b, 0x2e, 0x6e,
+	0x12, 0x1b, 0x0a, 0x09, 0x66, 0x69, 0x6c, 0x65, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x02, 0x20,
+	0x01, 0x28, 0x09, 0x52, 0x08, 0x66, 0x69, 0x6c, 0x65, 0x4e, 0x61, 0x6d, 0x65, 0x22, 0x1a, 0x0a,
+	0x18, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x65, 0x6c, 0x65, 0x74,
+	0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x5d, 0x0a, 0x1d, 0x53, 0x74, 0x72,
+	0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x56, 0x65, 0x72, 0x73, 0x69,
+	0x6f, 0x6e, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1f, 0x0a, 0x0b, 0x62, 0x75,
+	0x63, 0x6b, 0x65, 0x74, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52,
+	0x0a, 0x62, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1b, 0x0a, 0x09, 0x66,
+	0x69, 0x6c, 0x65, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08,
+	0x66, 0x69, 0x6c, 0x65, 0x4e, 0x61, 0x6d, 0x65, 0x22, 0x4b, 0x0a, 0x1e, 0x53, 0x74, 0x72, 0x65,
+	0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f,
+	0x6e, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x29, 0x0a, 0x08, 0x76, 0x65,
+	0x72, 0x73, 0x69, 0x6f, 0x6e, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x0d, 0x2e, 0x6e,
+	0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x46, 0x69, 0x6c, 0x65, 0x52, 0x08, 0x76, 0x65, 0x72,
+	0x73, 0x69, 0x6f, 0x6e, 0x73, 0x32, 0xb9, 0x02, 0x0a, 0x07, 0x4e, 0x6f, 0x64, 0x65, 0x41, 0x50,
+	0x49, 0x12, 0x4d, 0x0a, 0x0c, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x43, 0x72, 0x65, 0x61, 0x74,
+	0x65, 0x12, 0x1c, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x42, 0x75, 0x63, 0x6b,
+	0x65, 0x74, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a,
+	0x1d, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74,
+	0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00,
+	0x12, 0x47, 0x0a, 0x0a, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x56, 0x69, 0x65, 0x77, 0x12, 0x1a,
+	0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x56,
+	0x69, 0x65, 0x77, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1b, 0x2e, 0x6e, 0x6f, 0x64,
+	0x65, 0x61, 0x70, 0x69, 0x2e, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x56, 0x69, 0x65, 0x77, 0x52,
+	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x47, 0x0a, 0x0a, 0x42, 0x75, 0x63,
+	0x6b, 0x65, 0x74, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x1a, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70,
+	0x69, 0x2e, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x71, 0x75,
+	0x65, 0x73, 0x74, 0x1a, 0x1b, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x42, 0x75,
+	0x63, 0x6b, 0x65, 0x74, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65,
+	0x22, 0x00, 0x12, 0x4d, 0x0a, 0x0c, 0x42, 0x75, 0x63, 0x6b, 0x65, 0x74, 0x44, 0x65, 0x6c, 0x65,
+	0x74, 0x65, 0x12, 0x1c, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x42, 0x75, 0x63,
+	0x6b, 0x65, 0x74, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
+	0x1a, 0x1d, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x42, 0x75, 0x63, 0x6b, 0x65,
+	0x74, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22,
+	0x00, 0x32, 0x83, 0x0b, 0x0a, 0x09, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x41, 0x50, 0x49, 0x12,
+	0x65, 0x0a, 0x10, 0x46, 0x69, 0x6c, 0x65, 0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x72, 0x65,
+	0x61, 0x74, 0x65, 0x12, 0x26, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74,
+	0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x72,
+	0x65, 0x61, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x27, 0x2e, 0x6e, 0x6f,
+	0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65,
+	0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x52, 0x65, 0x73, 0x70,
+	0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x74, 0x0a, 0x15, 0x46, 0x69, 0x6c, 0x65, 0x55, 0x70,
+	0x6c, 0x6f, 0x61, 0x64, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x12,
+	0x2b, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d,
+	0x46, 0x69, 0x6c, 0x65, 0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x43,
+	0x72, 0x65, 0x61, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x2c, 0x2e, 0x6e,
 	0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c,
 	0x65, 0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x43, 0x72, 0x65, 0x61,
-	0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x2c, 0x2e, 0x6e, 0x6f, 0x64, 0x65,
-	0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x55, 0x70,
-	0x6c, 0x6f, 0x61, 0x64, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x52,
-	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x5b, 0x0a, 0x0f, 0x46, 0x69, 0x6c,
-	0x65, 0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x12, 0x1c, 0x2e, 0x6e,
-	0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c,
-	0x65, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x44, 0x61, 0x74, 0x61, 0x1a, 0x26, 0x2e, 0x6e, 0x6f, 0x64,
-	0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x55,
-	0x70, 0x6c, 0x6f, 0x61, 0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e,
-	0x73, 0x65, 0x22, 0x00, 0x28, 0x01, 0x12, 0x65, 0x0a, 0x10, 0x46, 0x69, 0x6c, 0x65, 0x55, 0x70,
-	0x6c, 0x6f, 0x61, 0x64, 0x43, 0x6f, 0x6d, 0x6d, 0x69, 0x74, 0x12, 0x26, 0x2e, 0x6e, 0x6f, 0x64,
-	0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x55,
-	0x70, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x6f, 0x6d, 0x6d, 0x69, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65,
-	0x73, 0x74, 0x1a, 0x27, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72,
-	0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x6f, 0x6d,
-	0x6d, 0x69, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x6b, 0x0a,
-	0x12, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x72, 0x65,
-	0x61, 0x74, 0x65, 0x12, 0x28, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74,
-	0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64,
-	0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x29, 0x2e,
+	0x74, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x5b, 0x0a, 0x0f,
+	0x46, 0x69, 0x6c, 0x65, 0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x12,
+	0x1c, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d,
+	0x46, 0x69, 0x6c, 0x65, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x44, 0x61, 0x74, 0x61, 0x1a, 0x26, 0x2e,
 	0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69,
-	0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65,
-	0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x75, 0x0a, 0x17, 0x46, 0x69,
-	0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x43,
-	0x72, 0x65, 0x61, 0x74, 0x65, 0x12, 0x2d, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e,
-	0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f,
-	0x61, 0x64, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x52, 0x65, 0x71,
-	0x75, 0x65, 0x73, 0x74, 0x1a, 0x29, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53,
-	0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61,
-	0x64, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22,
-	0x00, 0x12, 0x7a, 0x0a, 0x17, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61,
-	0x64, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x12, 0x2d, 0x2e, 0x6e,
-	0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c,
-	0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x43, 0x72,
-	0x65, 0x61, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x2e, 0x2e, 0x6e, 0x6f,
-	0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65,
-	0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x43, 0x72, 0x65,
-	0x61, 0x74, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x7e, 0x0a,
-	0x19, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x68, 0x75,
-	0x6e, 0x6b, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x56, 0x32, 0x12, 0x2d, 0x2e, 0x6e, 0x6f, 0x64,
-	0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44,
-	0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x43, 0x72, 0x65, 0x61,
-	0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x30, 0x2e, 0x6e, 0x6f, 0x64, 0x65,
-	0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f,
-	0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x43, 0x72, 0x65, 0x61, 0x74,
-	0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x56, 0x32, 0x22, 0x00, 0x12, 0x5e, 0x0a,
-	0x11, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x42, 0x6c, 0x6f,
-	0x63, 0x6b, 0x12, 0x27, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72,
-	0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x42,
-	0x6c, 0x6f, 0x63, 0x6b, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1c, 0x2e, 0x6e, 0x6f,
-	0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65,
-	0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x44, 0x61, 0x74, 0x61, 0x22, 0x00, 0x30, 0x01, 0x12, 0x4d, 0x0a,
-	0x08, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x1e, 0x2e, 0x6e, 0x6f, 0x64, 0x65,
-	0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69,
-	0x73, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1f, 0x2e, 0x6e, 0x6f, 0x64, 0x65,
-	0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69,
-	0x73, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x5f, 0x0a, 0x0e,
-	0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x73, 0x12, 0x24,
+	0x6c, 0x65, 0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x52, 0x65, 0x73,
+	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x28, 0x01, 0x12, 0x65, 0x0a, 0x10, 0x46, 0x69, 0x6c,
+	0x65, 0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x6f, 0x6d, 0x6d, 0x69, 0x74, 0x12, 0x26, 0x2e,
+	0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69,
+	0x6c, 0x65, 0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x6f, 0x6d, 0x6d, 0x69, 0x74, 0x52, 0x65,
+	0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x27, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e,
+	0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64,
+	0x43, 0x6f, 0x6d, 0x6d, 0x69, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00,
+	0x12, 0x6b, 0x0a, 0x12, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64,
+	0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x12, 0x28, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69,
+	0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c,
+	0x6f, 0x61, 0x64, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
+	0x1a, 0x29, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61,
+	0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x72, 0x65,
+	0x61, 0x74, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x75, 0x0a,
+	0x17, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x52, 0x61, 0x6e,
+	0x67, 0x65, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x12, 0x2d, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61,
+	0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77,
+	0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65,
+	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x29, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70,
+	0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e,
+	0x6c, 0x6f, 0x61, 0x64, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e,
+	0x73, 0x65, 0x22, 0x00, 0x12, 0x7a, 0x0a, 0x17, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e,
+	0x6c, 0x6f, 0x61, 0x64, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x12,
+	0x2d, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d,
+	0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x68, 0x75, 0x6e,
+	0x6b, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x2e,
 	0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46,
-	0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x73, 0x52, 0x65, 0x71,
-	0x75, 0x65, 0x73, 0x74, 0x1a, 0x25, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53,
-	0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x43, 0x68, 0x75,
-	0x6e, 0x6b, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x4d, 0x0a,
-	0x08, 0x46, 0x69, 0x6c, 0x65, 0x56, 0x69, 0x65, 0x77, 0x12, 0x1e, 0x2e, 0x6e, 0x6f, 0x64, 0x65,
-	0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x56, 0x69,
-	0x65, 0x77, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1f, 0x2e, 0x6e, 0x6f, 0x64, 0x65,
-	0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x56, 0x69,
-	0x65, 0x77, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x61, 0x0a, 0x0c,
-	0x46, 0x69, 0x6c, 0x65, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x73, 0x12, 0x26, 0x2e, 0x6e,
+	0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x43, 0x68, 0x75, 0x6e, 0x6b,
+	0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00,
+	0x12, 0x5e, 0x0a, 0x11, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f, 0x61, 0x64,
+	0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x12, 0x27, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e,
+	0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x6f, 0x77, 0x6e, 0x6c, 0x6f,
+	0x61, 0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1c,
+	0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46,
+	0x69, 0x6c, 0x65, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x44, 0x61, 0x74, 0x61, 0x22, 0x00, 0x30, 0x01,
+	0x12, 0x4d, 0x0a, 0x08, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x1e, 0x2e, 0x6e,
 	0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c,
-	0x65, 0x4c, 0x69, 0x73, 0x74, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x73, 0x52, 0x65, 0x71,
-	0x75, 0x65, 0x73, 0x74, 0x1a, 0x27, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53,
-	0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x56, 0x65, 0x72,
-	0x73, 0x69, 0x6f, 0x6e, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12,
-	0x53, 0x0a, 0x0a, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x12, 0x20, 0x2e,
+	0x65, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1f, 0x2e, 0x6e,
+	0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c,
+	0x65, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12,
+	0x5f, 0x0a, 0x0e, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x43, 0x68, 0x75, 0x6e, 0x6b,
+	0x73, 0x12, 0x24, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65,
+	0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x43, 0x68, 0x75, 0x6e, 0x6b, 0x73,
+	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x25, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70,
+	0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74,
+	0x43, 0x68, 0x75, 0x6e, 0x6b, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00,
+	0x12, 0x4d, 0x0a, 0x08, 0x46, 0x69, 0x6c, 0x65, 0x56, 0x69, 0x65, 0x77, 0x12, 0x1e, 0x2e, 0x6e,
+	0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c,
+	0x65, 0x56, 0x69, 0x65, 0x77, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1f, 0x2e, 0x6e,
+	0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c,
+	0x65, 0x56, 0x69, 0x65, 0x77, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12,
+	0x61, 0x0a, 0x0c, 0x46, 0x69, 0x6c, 0x65, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x73, 0x12,
+	0x26, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d,
+	0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x73,
+	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x27, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70,
+	0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x4c, 0x69, 0x73, 0x74,
+	0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65,
+	0x22, 0x00, 0x12, 0x53, 0x0a, 0x0a, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65,
+	0x12, 0x20, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61,
+	0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65,
+	0x73, 0x74, 0x1a, 0x21, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72,
+	0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x52, 0x65, 0x73,
+	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x12, 0x5e, 0x0a, 0x14, 0x46, 0x69, 0x6c, 0x65, 0x55,
+	0x70, 0x6c, 0x6f, 0x61, 0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x55, 0x6e, 0x61, 0x72, 0x79, 0x12,
+	0x1c, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d,
+	0x46, 0x69, 0x6c, 0x65, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x44, 0x61, 0x74, 0x61, 0x1a, 0x26, 0x2e,
 	0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69,
-	0x6c, 0x65, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a,
-	0x21, 0x2e, 0x6e, 0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d,
-	0x46, 0x69, 0x6c, 0x65, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e,
-	0x73, 0x65, 0x22, 0x00, 0x12, 0x5e, 0x0a, 0x14, 0x46, 0x69, 0x6c, 0x65, 0x55, 0x70, 0x6c, 0x6f,
-	0x61, 0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x55, 0x6e, 0x61, 0x72, 0x79, 0x12, 0x1c, 0x2e, 0x6e,
-	0x6f, 0x64, 0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c,
-	0x65, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x44, 0x61, 0x74, 0x61, 0x1a, 0x26, 0x2e, 0x6e, 0x6f, 0x64,
-	0x65, 0x61, 0x70, 0x69, 0x2e, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x46, 0x69, 0x6c, 0x65, 0x55,
-	0x70, 0x6c, 0x6f, 0x61, 0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e,
-	0x73, 0x65, 0x22, 0x00, 0x42, 0x1b, 0x5a, 0x19, 0x61, 0x6b, 0x61, 0x76, 0x65, 0x2e, 0x61, 0x69,
-	0x2f, 0x61, 0x6b, 0x61, 0x76, 0x65, 0x2f, 0x70, 0x72, 0x69, 0x76, 0x61, 0x74, 0x65, 0x2f, 0x70,
-	0x62, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x6c, 0x65, 0x55, 0x70, 0x6c, 0x6f, 0x61, 0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x52, 0x65, 0x73,
+	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x42, 0x1b, 0x5a, 0x19, 0x61, 0x6b, 0x61, 0x76, 0x65,
+	0x2e, 0x61, 0x69, 0x2f, 0x61, 0x6b, 0x61, 0x76, 0x65, 0x2f, 0x70, 0x72, 0x69, 0x76, 0x61, 0x74,
+	0x65, 0x2f, 0x70, 0x62, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -3120,119 +2861,110 @@ func file_nodeapi_proto_rawDescGZIP() []byte {
 	return file_nodeapi_proto_rawDescData
 }
 
-var file_nodeapi_proto_msgTypes = make([]protoimpl.MessageInfo, 44)
+var file_nodeapi_proto_msgTypes = make([]protoimpl.MessageInfo, 40)
 var file_nodeapi_proto_goTypes = []any{
-	(*BucketCreateRequest)(nil),                                            // 0: nodeapi.BucketCreateRequest
-	(*BucketCreateResponse)(nil),                                           // 1: nodeapi.BucketCreateResponse
-	(*BucketViewRequest)(nil),                                              // 2: nodeapi.BucketViewRequest
-	(*BucketViewResponse)(nil),                                             // 3: nodeapi.BucketViewResponse
-	(*BucketListRequest)(nil),                                              // 4: nodeapi.BucketListRequest
-	(*BucketListResponse)(nil),                                             // 5: nodeapi.BucketListResponse
-	(*BucketDeleteRequest)(nil),                                            // 6: nodeapi.BucketDeleteRequest
-	(*BucketDeleteResponse)(nil),                                           // 7: nodeapi.BucketDeleteResponse
-	(*FileBlockData)(nil),                                                  // 8: nodeapi.FileBlockData
-	(*StreamFileUploadCreateRequest)(nil),                                  // 9: nodeapi.StreamFileUploadCreateRequest
-	(*StreamFileUploadCreateResponse)(nil),                                 // 10: nodeapi.StreamFileUploadCreateResponse
-	(*Chunk)(nil),                                                          // 11: nodeapi.Chunk
-	(*StreamFileUploadChunkCreateRequest)(nil),                             // 12: nodeapi.StreamFileUploadChunkCreateRequest
-	(*StreamFileUploadChunkCreateResponse)(nil),                            // 13: nodeapi.StreamFileUploadChunkCreateResponse
-	(*StreamFileBlockData)(nil),                                            // 14: nodeapi.StreamFileBlockData
-	(*StreamFileUploadBlockResponse)(nil),                                  // 15: nodeapi.StreamFileUploadBlockResponse
-	(*StreamFileUploadCommitRequest)(nil),                                  // 16: nodeapi.StreamFileUploadCommitRequest
-	(*StreamFileUploadCommitResponse)(nil),                                 // 17: nodeapi.StreamFileUploadCommitResponse
-	(*StreamFileDownloadCreateRequest)(nil),                                // 18: nodeapi.StreamFileDownloadCreateRequest
-	(*StreamFileDownloadCreateResponse)(nil),                               // 19: nodeapi.StreamFileDownloadCreateResponse
-	(*StreamFileDownloadRangeCreateRequest)(nil),                           // 20: nodeapi.StreamFileDownloadRangeCreateRequest
-	(*StreamFileDownloadChunkCreateRequest)(nil),                           // 21: nodeapi.StreamFileDownloadChunkCreateRequest
-	(*StreamFileDownloadChunkCreateResponse)(nil),                          // 22: nodeapi.StreamFileDownloadChunkCreateResponse
-	(*StreamFileDownloadChunkCreateResponseV2)(nil),                        // 23: nodeapi.StreamFileDownloadChunkCreateResponseV2
-	(*StreamFileDownloadBlockRequest)(nil),                                 // 24: nodeapi.StreamFileDownloadBlockRequest
-	(*StreamFileListRequest)(nil),                                          // 25: nodeapi.StreamFileListRequest
-	(*File)(nil),                                                           // 26: nodeapi.File
-	(*StreamFileListResponse)(nil),                                         // 27: nodeapi.StreamFileListResponse
-	(*StreamFileListChunksRequest)(nil),                                    // 28: nodeapi.StreamFileListChunksRequest
-	(*StreamFileListChunksResponse)(nil),                                   // 29: nodeapi.StreamFileListChunksResponse
-	(*StreamFileViewRequest)(nil),                                          // 30: nodeapi.StreamFileViewRequest
-	(*StreamFileViewResponse)(nil),                                         // 31: nodeapi.StreamFileViewResponse
-	(*StreamFileDeleteRequest)(nil),                                        // 32: nodeapi.StreamFileDeleteRequest
-	(*StreamFileDeleteResponse)(nil),                                       // 33: nodeapi.StreamFileDeleteResponse
-	(*StreamFileListVersionsRequest)(nil),                                  // 34: nodeapi.StreamFileListVersionsRequest
-	(*StreamFileListVersionsResponse)(nil),                                 // 35: nodeapi.StreamFileListVersionsResponse
-	(*BucketListResponse_Bucket)(nil),                                      // 36: nodeapi.BucketListResponse.Bucket
-	(*Chunk_Block)(nil),                                                    // 37: nodeapi.Chunk.Block
-	(*StreamFileUploadChunkCreateResponse_BlockUpload)(nil),                // 38: nodeapi.StreamFileUploadChunkCreateResponse.BlockUpload
-	(*StreamFileDownloadCreateResponse_Chunk)(nil),                         // 39: nodeapi.StreamFileDownloadCreateResponse.Chunk
-	(*StreamFileDownloadChunkCreateResponse_BlockDownload)(nil),            // 40: nodeapi.StreamFileDownloadChunkCreateResponse.BlockDownload
-	(*StreamFileDownloadChunkCreateResponseV2_BlockDownload)(nil),          // 41: nodeapi.StreamFileDownloadChunkCreateResponseV2.BlockDownload
-	(*StreamFileDownloadChunkCreateResponseV2_BlockDownload_Akave)(nil),    // 42: nodeapi.StreamFileDownloadChunkCreateResponseV2.BlockDownload.Akave
-	(*StreamFileDownloadChunkCreateResponseV2_BlockDownload_Filecoin)(nil), // 43: nodeapi.StreamFileDownloadChunkCreateResponseV2.BlockDownload.Filecoin
-	(*timestamppb.Timestamp)(nil),                                          // 44: google.protobuf.Timestamp
+	(*BucketCreateRequest)(nil),                                 // 0: nodeapi.BucketCreateRequest
+	(*BucketCreateResponse)(nil),                                // 1: nodeapi.BucketCreateResponse
+	(*BucketViewRequest)(nil),                                   // 2: nodeapi.BucketViewRequest
+	(*BucketViewResponse)(nil),                                  // 3: nodeapi.BucketViewResponse
+	(*BucketListRequest)(nil),                                   // 4: nodeapi.BucketListRequest
+	(*BucketListResponse)(nil),                                  // 5: nodeapi.BucketListResponse
+	(*BucketDeleteRequest)(nil),                                 // 6: nodeapi.BucketDeleteRequest
+	(*BucketDeleteResponse)(nil),                                // 7: nodeapi.BucketDeleteResponse
+	(*FileBlockData)(nil),                                       // 8: nodeapi.FileBlockData
+	(*StreamFileUploadCreateRequest)(nil),                       // 9: nodeapi.StreamFileUploadCreateRequest
+	(*StreamFileUploadCreateResponse)(nil),                      // 10: nodeapi.StreamFileUploadCreateResponse
+	(*Chunk)(nil),                                               // 11: nodeapi.Chunk
+	(*StreamFileUploadChunkCreateRequest)(nil),                  // 12: nodeapi.StreamFileUploadChunkCreateRequest
+	(*StreamFileUploadChunkCreateResponse)(nil),                 // 13: nodeapi.StreamFileUploadChunkCreateResponse
+	(*StreamFileBlockData)(nil),                                 // 14: nodeapi.StreamFileBlockData
+	(*StreamFileUploadBlockResponse)(nil),                       // 15: nodeapi.StreamFileUploadBlockResponse
+	(*StreamFileUploadCommitRequest)(nil),                       // 16: nodeapi.StreamFileUploadCommitRequest
+	(*StreamFileUploadCommitResponse)(nil),                      // 17: nodeapi.StreamFileUploadCommitResponse
+	(*StreamFileDownloadCreateRequest)(nil),                     // 18: nodeapi.StreamFileDownloadCreateRequest
+	(*StreamFileDownloadCreateResponse)(nil),                    // 19: nodeapi.StreamFileDownloadCreateResponse
+	(*StreamFileDownloadRangeCreateRequest)(nil),                // 20: nodeapi.StreamFileDownloadRangeCreateRequest
+	(*StreamFileDownloadChunkCreateRequest)(nil),                // 21: nodeapi.StreamFileDownloadChunkCreateRequest
+	(*StreamFileDownloadChunkCreateResponse)(nil),               // 22: nodeapi.StreamFileDownloadChunkCreateResponse
+	(*StreamFileDownloadBlockRequest)(nil),                      // 23: nodeapi.StreamFileDownloadBlockRequest
+	(*StreamFileListRequest)(nil),                               // 24: nodeapi.StreamFileListRequest
+	(*File)(nil),                                                // 25: nodeapi.File
+	(*StreamFileListResponse)(nil),                              // 26: nodeapi.StreamFileListResponse
+	(*StreamFileListChunksRequest)(nil),                         // 27: nodeapi.StreamFileListChunksRequest
+	(*StreamFileListChunksResponse)(nil),                        // 28: nodeapi.StreamFileListChunksResponse
+	(*StreamFileViewRequest)(nil),                               // 29: nodeapi.StreamFileViewRequest
+	(*StreamFileViewResponse)(nil),                              // 30: nodeapi.StreamFileViewResponse
+	(*StreamFileDeleteRequest)(nil),                             // 31: nodeapi.StreamFileDeleteRequest
+	(*StreamFileDeleteResponse)(nil),                            // 32: nodeapi.StreamFileDeleteResponse
+	(*StreamFileListVersionsRequest)(nil),                       // 33: nodeapi.StreamFileListVersionsRequest
+	(*StreamFileListVersionsResponse)(nil),                      // 34: nodeapi.StreamFileListVersionsResponse
+	(*BucketListResponse_Bucket)(nil),                           // 35: nodeapi.BucketListResponse.Bucket
+	(*Chunk_Block)(nil),                                         // 36: nodeapi.Chunk.Block
+	(*StreamFileUploadChunkCreateResponse_BlockUpload)(nil),     // 37: nodeapi.StreamFileUploadChunkCreateResponse.BlockUpload
+	(*StreamFileDownloadCreateResponse_Chunk)(nil),              // 38: nodeapi.StreamFileDownloadCreateResponse.Chunk
+	(*StreamFileDownloadChunkCreateResponse_BlockDownload)(nil), // 39: nodeapi.StreamFileDownloadChunkCreateResponse.BlockDownload
+	(*timestamppb.Timestamp)(nil),                               // 40: google.protobuf.Timestamp
 }
 var file_nodeapi_proto_depIdxs = []int32{
-	44, // 0: nodeapi.BucketCreateResponse.created_at:type_name -> google.protobuf.Timestamp
-	44, // 1: nodeapi.BucketViewResponse.created_at:type_name -> google.protobuf.Timestamp
-	36, // 2: nodeapi.BucketListResponse.buckets:type_name -> nodeapi.BucketListResponse.Bucket
-	44, // 3: nodeapi.StreamFileUploadCreateResponse.created_at:type_name -> google.protobuf.Timestamp
-	37, // 4: nodeapi.Chunk.blocks:type_name -> nodeapi.Chunk.Block
+	40, // 0: nodeapi.BucketCreateResponse.created_at:type_name -> google.protobuf.Timestamp
+	40, // 1: nodeapi.BucketViewResponse.created_at:type_name -> google.protobuf.Timestamp
+	35, // 2: nodeapi.BucketListResponse.buckets:type_name -> nodeapi.BucketListResponse.Bucket
+	40, // 3: nodeapi.StreamFileUploadCreateResponse.created_at:type_name -> google.protobuf.Timestamp
+	36, // 4: nodeapi.Chunk.blocks:type_name -> nodeapi.Chunk.Block
 	11, // 5: nodeapi.StreamFileUploadChunkCreateRequest.chunk:type_name -> nodeapi.Chunk
-	38, // 6: nodeapi.StreamFileUploadChunkCreateResponse.blocks:type_name -> nodeapi.StreamFileUploadChunkCreateResponse.BlockUpload
+	37, // 6: nodeapi.StreamFileUploadChunkCreateResponse.blocks:type_name -> nodeapi.StreamFileUploadChunkCreateResponse.BlockUpload
 	11, // 7: nodeapi.StreamFileBlockData.chunk:type_name -> nodeapi.Chunk
-	44, // 8: nodeapi.StreamFileUploadCommitResponse.committed_at:type_name -> google.protobuf.Timestamp
-	39, // 9: nodeapi.StreamFileDownloadCreateResponse.chunks:type_name -> nodeapi.StreamFileDownloadCreateResponse.Chunk
-	40, // 10: nodeapi.StreamFileDownloadChunkCreateResponse.blocks:type_name -> nodeapi.StreamFileDownloadChunkCreateResponse.BlockDownload
-	41, // 11: nodeapi.StreamFileDownloadChunkCreateResponseV2.blocks:type_name -> nodeapi.StreamFileDownloadChunkCreateResponseV2.BlockDownload
-	44, // 12: nodeapi.File.created_at:type_name -> google.protobuf.Timestamp
-	44, // 13: nodeapi.File.commited_at:type_name -> google.protobuf.Timestamp
-	26, // 14: nodeapi.StreamFileListResponse.files:type_name -> nodeapi.File
-	44, // 15: nodeapi.StreamFileViewResponse.created_at:type_name -> google.protobuf.Timestamp
-	44, // 16: nodeapi.StreamFileViewResponse.committed_at:type_name -> google.protobuf.Timestamp
-	26, // 17: nodeapi.StreamFileListVersionsResponse.versions:type_name -> nodeapi.File
-	44, // 18: nodeapi.BucketListResponse.Bucket.created_at:type_name -> google.protobuf.Timestamp
-	42, // 19: nodeapi.StreamFileDownloadChunkCreateResponseV2.BlockDownload.akave:type_name -> nodeapi.StreamFileDownloadChunkCreateResponseV2.BlockDownload.Akave
-	43, // 20: nodeapi.StreamFileDownloadChunkCreateResponseV2.BlockDownload.filecoin:type_name -> nodeapi.StreamFileDownloadChunkCreateResponseV2.BlockDownload.Filecoin
-	0,  // 21: nodeapi.NodeAPI.BucketCreate:input_type -> nodeapi.BucketCreateRequest
-	2,  // 22: nodeapi.NodeAPI.BucketView:input_type -> nodeapi.BucketViewRequest
-	4,  // 23: nodeapi.NodeAPI.BucketList:input_type -> nodeapi.BucketListRequest
-	6,  // 24: nodeapi.NodeAPI.BucketDelete:input_type -> nodeapi.BucketDeleteRequest
-	9,  // 25: nodeapi.StreamAPI.FileUploadCreate:input_type -> nodeapi.StreamFileUploadCreateRequest
-	12, // 26: nodeapi.StreamAPI.FileUploadChunkCreate:input_type -> nodeapi.StreamFileUploadChunkCreateRequest
-	14, // 27: nodeapi.StreamAPI.FileUploadBlock:input_type -> nodeapi.StreamFileBlockData
-	16, // 28: nodeapi.StreamAPI.FileUploadCommit:input_type -> nodeapi.StreamFileUploadCommitRequest
-	18, // 29: nodeapi.StreamAPI.FileDownloadCreate:input_type -> nodeapi.StreamFileDownloadCreateRequest
-	20, // 30: nodeapi.StreamAPI.FileDownloadRangeCreate:input_type -> nodeapi.StreamFileDownloadRangeCreateRequest
-	21, // 31: nodeapi.StreamAPI.FileDownloadChunkCreate:input_type -> nodeapi.StreamFileDownloadChunkCreateRequest
-	21, // 32: nodeapi.StreamAPI.FileDownloadChunkCreateV2:input_type -> nodeapi.StreamFileDownloadChunkCreateRequest
-	24, // 33: nodeapi.StreamAPI.FileDownloadBlock:input_type -> nodeapi.StreamFileDownloadBlockRequest
-	25, // 34: nodeapi.StreamAPI.FileList:input_type -> nodeapi.StreamFileListRequest
-	28, // 35: nodeapi.StreamAPI.FileListChunks:input_type -> nodeapi.StreamFileListChunksRequest
-	30, // 36: nodeapi.StreamAPI.FileView:input_type -> nodeapi.StreamFileViewRequest
-	34, // 37: nodeapi.StreamAPI.FileVersions:input_type -> nodeapi.StreamFileListVersionsRequest
-	32, // 38: nodeapi.StreamAPI.FileDelete:input_type -> nodeapi.StreamFileDeleteRequest
-	14, // 39: nodeapi.StreamAPI.FileUploadBlockUnary:input_type -> nodeapi.StreamFileBlockData
-	1,  // 40: nodeapi.NodeAPI.BucketCreate:output_type -> nodeapi.BucketCreateResponse
-	3,  // 41: nodeapi.NodeAPI.BucketView:output_type -> nodeapi.BucketViewResponse
-	5,  // 42: nodeapi.NodeAPI.BucketList:output_type -> nodeapi.BucketListResponse
-	7,  // 43: nodeapi.NodeAPI.BucketDelete:output_type -> nodeapi.BucketDeleteResponse
-	10, // 44: nodeapi.StreamAPI.FileUploadCreate:output_type -> nodeapi.StreamFileUploadCreateResponse
-	13, // 45: nodeapi.StreamAPI.FileUploadChunkCreate:output_type -> nodeapi.StreamFileUploadChunkCreateResponse
-	15, // 46: nodeapi.StreamAPI.FileUploadBlock:output_type -> nodeapi.StreamFileUploadBlockResponse
-	17, // 47: nodeapi.StreamAPI.FileUploadCommit:output_type -> nodeapi.StreamFileUploadCommitResponse
-	19, // 48: nodeapi.StreamAPI.FileDownloadCreate:output_type -> nodeapi.StreamFileDownloadCreateResponse
-	19, // 49: nodeapi.StreamAPI.FileDownloadRangeCreate:output_type -> nodeapi.StreamFileDownloadCreateResponse
-	22, // 50: nodeapi.StreamAPI.FileDownloadChunkCreate:output_type -> nodeapi.StreamFileDownloadChunkCreateResponse
-	23, // 51: nodeapi.StreamAPI.FileDownloadChunkCreateV2:output_type -> nodeapi.StreamFileDownloadChunkCreateResponseV2
-	14, // 52: nodeapi.StreamAPI.FileDownloadBlock:output_type -> nodeapi.StreamFileBlockData
-	27, // 53: nodeapi.StreamAPI.FileList:output_type -> nodeapi.StreamFileListResponse
-	29, // 54: nodeapi.StreamAPI.FileListChunks:output_type -> nodeapi.StreamFileListChunksResponse
-	31, // 55: nodeapi.StreamAPI.FileView:output_type -> nodeapi.StreamFileViewResponse
-	35, // 56: nodeapi.StreamAPI.FileVersions:output_type -> nodeapi.StreamFileListVersionsResponse
-	33, // 57: nodeapi.StreamAPI.FileDelete:output_type -> nodeapi.StreamFileDeleteResponse
-	15, // 58: nodeapi.StreamAPI.FileUploadBlockUnary:output_type -> nodeapi.StreamFileUploadBlockResponse
-	40, // [40:59] is the sub-list for method output_type
-	21, // [21:40] is the sub-list for method input_type
-	21, // [21:21] is the sub-list for extension type_name
-	21, // [21:21] is the sub-list for extension extendee
-	0,  // [0:21] is the sub-list for field type_name
+	40, // 8: nodeapi.StreamFileUploadCommitResponse.committed_at:type_name -> google.protobuf.Timestamp
+	38, // 9: nodeapi.StreamFileDownloadCreateResponse.chunks:type_name -> nodeapi.StreamFileDownloadCreateResponse.Chunk
+	39, // 10: nodeapi.StreamFileDownloadChunkCreateResponse.blocks:type_name -> nodeapi.StreamFileDownloadChunkCreateResponse.BlockDownload
+	40, // 11: nodeapi.File.created_at:type_name -> google.protobuf.Timestamp
+	40, // 12: nodeapi.File.commited_at:type_name -> google.protobuf.Timestamp
+	25, // 13: nodeapi.StreamFileListResponse.files:type_name -> nodeapi.File
+	40, // 14: nodeapi.StreamFileViewResponse.created_at:type_name -> google.protobuf.Timestamp
+	40, // 15: nodeapi.StreamFileViewResponse.committed_at:type_name -> google.protobuf.Timestamp
+	25, // 16: nodeapi.StreamFileListVersionsResponse.versions:type_name -> nodeapi.File
+	40, // 17: nodeapi.BucketListResponse.Bucket.created_at:type_name -> google.protobuf.Timestamp
+	0,  // 18: nodeapi.NodeAPI.BucketCreate:input_type -> nodeapi.BucketCreateRequest
+	2,  // 19: nodeapi.NodeAPI.BucketView:input_type -> nodeapi.BucketViewRequest
+	4,  // 20: nodeapi.NodeAPI.BucketList:input_type -> nodeapi.BucketListRequest
+	6,  // 21: nodeapi.NodeAPI.BucketDelete:input_type -> nodeapi.BucketDeleteRequest
+	9,  // 22: nodeapi.StreamAPI.FileUploadCreate:input_type -> nodeapi.StreamFileUploadCreateRequest
+	12, // 23: nodeapi.StreamAPI.FileUploadChunkCreate:input_type -> nodeapi.StreamFileUploadChunkCreateRequest
+	14, // 24: nodeapi.StreamAPI.FileUploadBlock:input_type -> nodeapi.StreamFileBlockData
+	16, // 25: nodeapi.StreamAPI.FileUploadCommit:input_type -> nodeapi.StreamFileUploadCommitRequest
+	18, // 26: nodeapi.StreamAPI.FileDownloadCreate:input_type -> nodeapi.StreamFileDownloadCreateRequest
+	20, // 27: nodeapi.StreamAPI.FileDownloadRangeCreate:input_type -> nodeapi.StreamFileDownloadRangeCreateRequest
+	21, // 28: nodeapi.StreamAPI.FileDownloadChunkCreate:input_type -> nodeapi.StreamFileDownloadChunkCreateRequest
+	23, // 29: nodeapi.StreamAPI.FileDownloadBlock:input_type -> nodeapi.StreamFileDownloadBlockRequest
+	24, // 30: nodeapi.StreamAPI.FileList:input_type -> nodeapi.StreamFileListRequest
+	27, // 31: nodeapi.StreamAPI.FileListChunks:input_type -> nodeapi.StreamFileListChunksRequest
+	29, // 32: nodeapi.StreamAPI.FileView:input_type -> nodeapi.StreamFileViewRequest
+	33, // 33: nodeapi.StreamAPI.FileVersions:input_type -> nodeapi.StreamFileListVersionsRequest
+	31, // 34: nodeapi.StreamAPI.FileDelete:input_type -> nodeapi.StreamFileDeleteRequest
+	14, // 35: nodeapi.StreamAPI.FileUploadBlockUnary:input_type -> nodeapi.StreamFileBlockData
+	1,  // 36: nodeapi.NodeAPI.BucketCreate:output_type -> nodeapi.BucketCreateResponse
+	3,  // 37: nodeapi.NodeAPI.BucketView:output_type -> nodeapi.BucketViewResponse
+	5,  // 38: nodeapi.NodeAPI.BucketList:output_type -> nodeapi.BucketListResponse
+	7,  // 39: nodeapi.NodeAPI.BucketDelete:output_type -> nodeapi.BucketDeleteResponse
+	10, // 40: nodeapi.StreamAPI.FileUploadCreate:output_type -> nodeapi.StreamFileUploadCreateResponse
+	13, // 41: nodeapi.StreamAPI.FileUploadChunkCreate:output_type -> nodeapi.StreamFileUploadChunkCreateResponse
+	15, // 42: nodeapi.StreamAPI.FileUploadBlock:output_type -> nodeapi.StreamFileUploadBlockResponse
+	17, // 43: nodeapi.StreamAPI.FileUploadCommit:output_type -> nodeapi.StreamFileUploadCommitResponse
+	19, // 44: nodeapi.StreamAPI.FileDownloadCreate:output_type -> nodeapi.StreamFileDownloadCreateResponse
+	19, // 45: nodeapi.StreamAPI.FileDownloadRangeCreate:output_type -> nodeapi.StreamFileDownloadCreateResponse
+	22, // 46: nodeapi.StreamAPI.FileDownloadChunkCreate:output_type -> nodeapi.StreamFileDownloadChunkCreateResponse
+	14, // 47: nodeapi.StreamAPI.FileDownloadBlock:output_type -> nodeapi.StreamFileBlockData
+	26, // 48: nodeapi.StreamAPI.FileList:output_type -> nodeapi.StreamFileListResponse
+	28, // 49: nodeapi.StreamAPI.FileListChunks:output_type -> nodeapi.StreamFileListChunksResponse
+	30, // 50: nodeapi.StreamAPI.FileView:output_type -> nodeapi.StreamFileViewResponse
+	34, // 51: nodeapi.StreamAPI.FileVersions:output_type -> nodeapi.StreamFileListVersionsResponse
+	32, // 52: nodeapi.StreamAPI.FileDelete:output_type -> nodeapi.StreamFileDeleteResponse
+	15, // 53: nodeapi.StreamAPI.FileUploadBlockUnary:output_type -> nodeapi.StreamFileUploadBlockResponse
+	36, // [36:54] is the sub-list for method output_type
+	18, // [18:36] is the sub-list for method input_type
+	18, // [18:18] is the sub-list for extension type_name
+	18, // [18:18] is the sub-list for extension extendee
+	0,  // [0:18] is the sub-list for field type_name
 }
 
 func init() { file_nodeapi_proto_init() }
@@ -3518,18 +3250,6 @@ func file_nodeapi_proto_init() {
 			}
 		}
 		file_nodeapi_proto_msgTypes[23].Exporter = func(v any, i int) any {
-			switch v := v.(*StreamFileDownloadChunkCreateResponseV2); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_nodeapi_proto_msgTypes[24].Exporter = func(v any, i int) any {
 			switch v := v.(*StreamFileDownloadBlockRequest); i {
 			case 0:
 				return &v.state
@@ -3541,7 +3261,7 @@ func file_nodeapi_proto_init() {
 				return nil
 			}
 		}
-		file_nodeapi_proto_msgTypes[25].Exporter = func(v any, i int) any {
+		file_nodeapi_proto_msgTypes[24].Exporter = func(v any, i int) any {
 			switch v := v.(*StreamFileListRequest); i {
 			case 0:
 				return &v.state
@@ -3553,7 +3273,7 @@ func file_nodeapi_proto_init() {
 				return nil
 			}
 		}
-		file_nodeapi_proto_msgTypes[26].Exporter = func(v any, i int) any {
+		file_nodeapi_proto_msgTypes[25].Exporter = func(v any, i int) any {
 			switch v := v.(*File); i {
 			case 0:
 				return &v.state
@@ -3565,7 +3285,7 @@ func file_nodeapi_proto_init() {
 				return nil
 			}
 		}
-		file_nodeapi_proto_msgTypes[27].Exporter = func(v any, i int) any {
+		file_nodeapi_proto_msgTypes[26].Exporter = func(v any, i int) any {
 			switch v := v.(*StreamFileListResponse); i {
 			case 0:
 				return &v.state
@@ -3577,7 +3297,7 @@ func file_nodeapi_proto_init() {
 				return nil
 			}
 		}
-		file_nodeapi_proto_msgTypes[28].Exporter = func(v any, i int) any {
+		file_nodeapi_proto_msgTypes[27].Exporter = func(v any, i int) any {
 			switch v := v.(*StreamFileListChunksRequest); i {
 			case 0:
 				return &v.state
@@ -3589,7 +3309,7 @@ func file_nodeapi_proto_init() {
 				return nil
 			}
 		}
-		file_nodeapi_proto_msgTypes[29].Exporter = func(v any, i int) any {
+		file_nodeapi_proto_msgTypes[28].Exporter = func(v any, i int) any {
 			switch v := v.(*StreamFileListChunksResponse); i {
 			case 0:
 				return &v.state
@@ -3601,7 +3321,7 @@ func file_nodeapi_proto_init() {
 				return nil
 			}
 		}
-		file_nodeapi_proto_msgTypes[30].Exporter = func(v any, i int) any {
+		file_nodeapi_proto_msgTypes[29].Exporter = func(v any, i int) any {
 			switch v := v.(*StreamFileViewRequest); i {
 			case 0:
 				return &v.state
@@ -3613,7 +3333,7 @@ func file_nodeapi_proto_init() {
 				return nil
 			}
 		}
-		file_nodeapi_proto_msgTypes[31].Exporter = func(v any, i int) any {
+		file_nodeapi_proto_msgTypes[30].Exporter = func(v any, i int) any {
 			switch v := v.(*StreamFileViewResponse); i {
 			case 0:
 				return &v.state
@@ -3625,7 +3345,7 @@ func file_nodeapi_proto_init() {
 				return nil
 			}
 		}
-		file_nodeapi_proto_msgTypes[32].Exporter = func(v any, i int) any {
+		file_nodeapi_proto_msgTypes[31].Exporter = func(v any, i int) any {
 			switch v := v.(*StreamFileDeleteRequest); i {
 			case 0:
 				return &v.state
@@ -3637,7 +3357,7 @@ func file_nodeapi_proto_init() {
 				return nil
 			}
 		}
-		file_nodeapi_proto_msgTypes[33].Exporter = func(v any, i int) any {
+		file_nodeapi_proto_msgTypes[32].Exporter = func(v any, i int) any {
 			switch v := v.(*StreamFileDeleteResponse); i {
 			case 0:
 				return &v.state
@@ -3649,7 +3369,7 @@ func file_nodeapi_proto_init() {
 				return nil
 			}
 		}
-		file_nodeapi_proto_msgTypes[34].Exporter = func(v any, i int) any {
+		file_nodeapi_proto_msgTypes[33].Exporter = func(v any, i int) any {
 			switch v := v.(*StreamFileListVersionsRequest); i {
 			case 0:
 				return &v.state
@@ -3661,7 +3381,7 @@ func file_nodeapi_proto_init() {
 				return nil
 			}
 		}
-		file_nodeapi_proto_msgTypes[35].Exporter = func(v any, i int) any {
+		file_nodeapi_proto_msgTypes[34].Exporter = func(v any, i int) any {
 			switch v := v.(*StreamFileListVersionsResponse); i {
 			case 0:
 				return &v.state
@@ -3673,7 +3393,7 @@ func file_nodeapi_proto_init() {
 				return nil
 			}
 		}
-		file_nodeapi_proto_msgTypes[36].Exporter = func(v any, i int) any {
+		file_nodeapi_proto_msgTypes[35].Exporter = func(v any, i int) any {
 			switch v := v.(*BucketListResponse_Bucket); i {
 			case 0:
 				return &v.state
@@ -3685,7 +3405,7 @@ func file_nodeapi_proto_init() {
 				return nil
 			}
 		}
-		file_nodeapi_proto_msgTypes[37].Exporter = func(v any, i int) any {
+		file_nodeapi_proto_msgTypes[36].Exporter = func(v any, i int) any {
 			switch v := v.(*Chunk_Block); i {
 			case 0:
 				return &v.state
@@ -3697,7 +3417,7 @@ func file_nodeapi_proto_init() {
 				return nil
 			}
 		}
-		file_nodeapi_proto_msgTypes[38].Exporter = func(v any, i int) any {
+		file_nodeapi_proto_msgTypes[37].Exporter = func(v any, i int) any {
 			switch v := v.(*StreamFileUploadChunkCreateResponse_BlockUpload); i {
 			case 0:
 				return &v.state
@@ -3709,7 +3429,7 @@ func file_nodeapi_proto_init() {
 				return nil
 			}
 		}
-		file_nodeapi_proto_msgTypes[39].Exporter = func(v any, i int) any {
+		file_nodeapi_proto_msgTypes[38].Exporter = func(v any, i int) any {
 			switch v := v.(*StreamFileDownloadCreateResponse_Chunk); i {
 			case 0:
 				return &v.state
@@ -3721,44 +3441,8 @@ func file_nodeapi_proto_init() {
 				return nil
 			}
 		}
-		file_nodeapi_proto_msgTypes[40].Exporter = func(v any, i int) any {
+		file_nodeapi_proto_msgTypes[39].Exporter = func(v any, i int) any {
 			switch v := v.(*StreamFileDownloadChunkCreateResponse_BlockDownload); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_nodeapi_proto_msgTypes[41].Exporter = func(v any, i int) any {
-			switch v := v.(*StreamFileDownloadChunkCreateResponseV2_BlockDownload); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_nodeapi_proto_msgTypes[42].Exporter = func(v any, i int) any {
-			switch v := v.(*StreamFileDownloadChunkCreateResponseV2_BlockDownload_Akave); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_nodeapi_proto_msgTypes[43].Exporter = func(v any, i int) any {
-			switch v := v.(*StreamFileDownloadChunkCreateResponseV2_BlockDownload_Filecoin); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -3776,7 +3460,7 @@ func file_nodeapi_proto_init() {
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_nodeapi_proto_rawDesc,
 			NumEnums:      0,
-			NumMessages:   44,
+			NumMessages:   40,
 			NumExtensions: 0,
 			NumServices:   2,
 		},
